Move inner classes around
