+++ group-coordinator/src/test/java/org/apache/kafka/coordinator/group/GroupMetadataManagerTest.java
@@ -50,76 +50,41 @@ import org.apache.kafka.common.message.SyncGroupRequestData;
 import org.apache.kafka.common.message.SyncGroupRequestData.SyncGroupRequestAssignment;
 import org.apache.kafka.common.message.SyncGroupResponseData;
 import org.apache.kafka.common.metadata.PartitionRecord;
-import org.apache.kafka.common.metadata.RegisterBrokerRecord;
 import org.apache.kafka.common.metadata.RemoveTopicRecord;
 import org.apache.kafka.common.metadata.TopicRecord;
-import org.apache.kafka.common.network.ClientInformation;
-import org.apache.kafka.common.network.ListenerName;
-import org.apache.kafka.common.protocol.ApiKeys;
-import org.apache.kafka.common.protocol.ApiMessage;
 import org.apache.kafka.common.protocol.Errors;
-import org.apache.kafka.common.requests.RequestContext;
-import org.apache.kafka.common.requests.RequestHeader;
-import org.apache.kafka.common.security.auth.KafkaPrincipal;
-import org.apache.kafka.common.security.auth.SecurityProtocol;
 import org.apache.kafka.common.utils.LogContext;
-import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.coordinator.group.MockCoordinatorTimer.ExpiredTimeout;
 import org.apache.kafka.coordinator.group.MockCoordinatorTimer.ScheduledTimeout;
-import org.apache.kafka.coordinator.group.assignor.AssignmentSpec;
 import org.apache.kafka.coordinator.group.assignor.GroupAssignment;
 import org.apache.kafka.coordinator.group.assignor.MemberAssignment;
 import org.apache.kafka.coordinator.group.assignor.PartitionAssignor;
 import org.apache.kafka.coordinator.group.assignor.PartitionAssignorException;
-import org.apache.kafka.coordinator.group.assignor.SubscribedTopicDescriber;
 import org.apache.kafka.coordinator.group.consumer.Assignment;
 import org.apache.kafka.coordinator.group.consumer.ConsumerGroup;
 import org.apache.kafka.coordinator.group.consumer.ConsumerGroupMember;
 import org.apache.kafka.coordinator.group.consumer.TopicMetadata;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupCurrentMemberAssignmentKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupCurrentMemberAssignmentValue;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupMemberMetadataKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupMemberMetadataValue;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupMetadataKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupMetadataValue;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupPartitionMetadataKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupPartitionMetadataValue;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupTargetAssignmentMemberKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupTargetAssignmentMemberValue;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupTargetAssignmentMetadataKey;
-import org.apache.kafka.coordinator.group.generated.ConsumerGroupTargetAssignmentMetadataValue;
-import org.apache.kafka.coordinator.group.generated.GroupMetadataKey;
 import org.apache.kafka.coordinator.group.generated.GroupMetadataValue;
 import org.apache.kafka.coordinator.group.classic.ClassicGroup;
 import org.apache.kafka.coordinator.group.classic.ClassicGroupMember;
-import org.apache.kafka.coordinator.group.classic.ClassicGroupState;
 import org.apache.kafka.coordinator.group.metrics.GroupCoordinatorMetricsShard;
 import org.apache.kafka.coordinator.group.runtime.CoordinatorResult;
 import org.apache.kafka.image.MetadataDelta;
 import org.apache.kafka.image.MetadataImage;
 import org.apache.kafka.image.MetadataProvenance;
-import org.apache.kafka.image.TopicImage;
-import org.apache.kafka.image.TopicsImage;
-import org.apache.kafka.server.common.ApiMessageAndVersion;
 import org.apache.kafka.server.common.MetadataVersion;
-import org.apache.kafka.timeline.SnapshotRegistry;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.opentest4j.AssertionFailedError;
 
-import java.net.InetAddress;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -148,7 +113,6 @@ import static org.apache.kafka.coordinator.group.classic.ClassicGroupState.PREPA
 import static org.apache.kafka.coordinator.group.classic.ClassicGroupState.STABLE;
 import static org.apache.kafka.coordinator.group.metrics.GroupCoordinatorMetrics.CLASSIC_GROUP_COMPLETED_REBALANCES_SENSOR_NAME;
 import static org.apache.kafka.coordinator.group.metrics.GroupCoordinatorMetrics.CONSUMER_GROUP_REBALANCES_SENSOR_NAME;
-import static org.junit.jupiter.api.AssertionFailureBuilder.assertionFailure;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -164,1093 +128,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class GroupMetadataManagerTest {
-    static class MockPartitionAssignor implements PartitionAssignor {
-        private final String name;
-        private GroupAssignment prepareGroupAssignment = null;
-
-        MockPartitionAssignor(String name) {
-            this.name = name;
-        }
-
-        public void prepareGroupAssignment(GroupAssignment prepareGroupAssignment) {
-            this.prepareGroupAssignment = prepareGroupAssignment;
-        }
-
-        @Override
-        public String name() {
-            return name;
-        }
-
-        @Override
-        public GroupAssignment assign(AssignmentSpec assignmentSpec, SubscribedTopicDescriber subscribedTopicDescriber) throws PartitionAssignorException {
-            return prepareGroupAssignment;
-        }
-    }
-
-    public static class MetadataImageBuilder {
-        private MetadataDelta delta = new MetadataDelta(MetadataImage.EMPTY);
-
-        public MetadataImageBuilder addTopic(
-            Uuid topicId,
-            String topicName,
-            int numPartitions
-        ) {
-            // For testing purposes, the following criteria are used:
-            // - Number of replicas for each partition: 2
-            // - Number of brokers available in the cluster: 4
-            delta.replay(new TopicRecord().setTopicId(topicId).setName(topicName));
-            for (int i = 0; i < numPartitions; i++) {
-                delta.replay(new PartitionRecord()
-                    .setTopicId(topicId)
-                    .setPartitionId(i)
-                    .setReplicas(Arrays.asList(i % 4, (i + 1) % 4)));
-            }
-            return this;
-        }
-
-        /**
-         * Add rack Ids for 4 broker Ids.
-         *
-         * For testing purposes, each broker is mapped
-         * to a rack Id with the same broker Id as a suffix.
-         */
-        public MetadataImageBuilder addRacks() {
-            for (int i = 0; i < 4; i++) {
-                delta.replay(new RegisterBrokerRecord().setBrokerId(i).setRack("rack" + i));
-            }
-            return this;
-        }
-
-        public MetadataImage build() {
-            return delta.apply(MetadataProvenance.EMPTY);
-        }
-    }
-
-    static class ConsumerGroupBuilder {
-        private final String groupId;
-        private final int groupEpoch;
-        private int assignmentEpoch;
-        private final Map<String, ConsumerGroupMember> members = new HashMap<>();
-        private final Map<String, Assignment> assignments = new HashMap<>();
-        private Map<String, TopicMetadata> subscriptionMetadata;
-
-        public ConsumerGroupBuilder(String groupId, int groupEpoch) {
-            this.groupId = groupId;
-            this.groupEpoch = groupEpoch;
-            this.assignmentEpoch = 0;
-        }
-
-        public ConsumerGroupBuilder withMember(ConsumerGroupMember member) {
-            this.members.put(member.memberId(), member);
-            return this;
-        }
-
-        public ConsumerGroupBuilder withSubscriptionMetadata(Map<String, TopicMetadata> subscriptionMetadata) {
-            this.subscriptionMetadata = subscriptionMetadata;
-            return this;
-        }
-
-        public ConsumerGroupBuilder withAssignment(String memberId, Map<Uuid, Set<Integer>> assignment) {
-            this.assignments.put(memberId, new Assignment(assignment));
-            return this;
-        }
-
-        public ConsumerGroupBuilder withAssignmentEpoch(int assignmentEpoch) {
-            this.assignmentEpoch = assignmentEpoch;
-            return this;
-        }
-
-        public List<Record> build(TopicsImage topicsImage) {
-            List<Record> records = new ArrayList<>();
-
-            // Add subscription records for members.
-            members.forEach((memberId, member) -> {
-                records.add(RecordHelpers.newMemberSubscriptionRecord(groupId, member));
-            });
-
-            // Add subscription metadata.
-            if (subscriptionMetadata == null) {
-                subscriptionMetadata = new HashMap<>();
-                members.forEach((memberId, member) -> {
-                    member.subscribedTopicNames().forEach(topicName -> {
-                        TopicImage topicImage = topicsImage.getTopic(topicName);
-                        if (topicImage != null) {
-                            subscriptionMetadata.put(topicName, new TopicMetadata(
-                                topicImage.id(),
-                                topicImage.name(),
-                                topicImage.partitions().size(),
-                                Collections.emptyMap()));
-                        }
-                    });
-                });
-            }
-
-            if (!subscriptionMetadata.isEmpty()) {
-                records.add(RecordHelpers.newGroupSubscriptionMetadataRecord(groupId, subscriptionMetadata));
-            }
-
-            // Add group epoch record.
-            records.add(RecordHelpers.newGroupEpochRecord(groupId, groupEpoch));
-
-            // Add target assignment records.
-            assignments.forEach((memberId, assignment) -> {
-                records.add(RecordHelpers.newTargetAssignmentRecord(groupId, memberId, assignment.partitions()));
-            });
-
-            // Add target assignment epoch.
-            records.add(RecordHelpers.newTargetAssignmentEpochRecord(groupId, assignmentEpoch));
-
-            // Add current assignment records for members.
-            members.forEach((memberId, member) -> {
-                records.add(RecordHelpers.newCurrentAssignmentRecord(groupId, member));
-            });
-
-            return records;
-        }
-    }
-
-    static class GroupMetadataManagerTestContext {
-        static class Builder {
-            final private MockTime time = new MockTime();
-            final private MockCoordinatorTimer<Void, Record> timer = new MockCoordinatorTimer<>(time);
-            final private LogContext logContext = new LogContext();
-            final private SnapshotRegistry snapshotRegistry = new SnapshotRegistry(logContext);
-            final private TopicPartition groupMetadataTopicPartition = new TopicPartition("topic", 0);
-            private MetadataImage metadataImage;
-            private List<PartitionAssignor> consumerGroupAssignors = Collections.singletonList(new MockPartitionAssignor("range"));
-            private List<ConsumerGroupBuilder> consumerGroupBuilders = new ArrayList<>();
-            private int consumerGroupMaxSize = Integer.MAX_VALUE;
-            private int consumerGroupMetadataRefreshIntervalMs = Integer.MAX_VALUE;
-            private int classicGroupMaxSize = Integer.MAX_VALUE;
-            private int classicGroupInitialRebalanceDelayMs = 3000;
-            final private int classicGroupNewMemberJoinTimeoutMs = 5 * 60 * 1000;
-            private int classicGroupMinSessionTimeoutMs = 10;
-            private int classicGroupMaxSessionTimeoutMs = 10 * 60 * 1000;
-            private GroupCoordinatorMetricsShard metrics = mock(GroupCoordinatorMetricsShard.class);
-
-            public Builder withMetadataImage(MetadataImage metadataImage) {
-                this.metadataImage = metadataImage;
-                return this;
-            }
-
-            public Builder withAssignors(List<PartitionAssignor> assignors) {
-                this.consumerGroupAssignors = assignors;
-                return this;
-            }
-
-            public Builder withConsumerGroup(ConsumerGroupBuilder builder) {
-                this.consumerGroupBuilders.add(builder);
-                return this;
-            }
-
-            public Builder withConsumerGroupMaxSize(int consumerGroupMaxSize) {
-                this.consumerGroupMaxSize = consumerGroupMaxSize;
-                return this;
-            }
-
-            public Builder withConsumerGroupMetadataRefreshIntervalMs(int consumerGroupMetadataRefreshIntervalMs) {
-                this.consumerGroupMetadataRefreshIntervalMs = consumerGroupMetadataRefreshIntervalMs;
-                return this;
-            }
-
-            public Builder withClassicGroupMaxSize(int classicGroupMaxSize) {
-                this.classicGroupMaxSize = classicGroupMaxSize;
-                return this;
-            }
-
-            public Builder withClassicGroupInitialRebalanceDelayMs(int classicGroupInitialRebalanceDelayMs) {
-                this.classicGroupInitialRebalanceDelayMs = classicGroupInitialRebalanceDelayMs;
-                return this;
-            }
-
-            public Builder withClassicGroupMinSessionTimeoutMs(int classicGroupMinSessionTimeoutMs) {
-                this.classicGroupMinSessionTimeoutMs = classicGroupMinSessionTimeoutMs;
-                return this;
-            }
-
-            public Builder withClassicGroupMaxSessionTimeoutMs(int classicGroupMaxSessionTimeoutMs) {
-                this.classicGroupMaxSessionTimeoutMs = classicGroupMaxSessionTimeoutMs;
-                return this;
-            }
-
-            public GroupMetadataManagerTestContext build() {
-                if (metadataImage == null) metadataImage = MetadataImage.EMPTY;
-                if (consumerGroupAssignors == null) consumerGroupAssignors = Collections.emptyList();
-
-                GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext(
-                    time,
-                    timer,
-                    snapshotRegistry,
-                    metrics,
-                    new GroupMetadataManager.Builder()
-                        .withSnapshotRegistry(snapshotRegistry)
-                        .withLogContext(logContext)
-                        .withTime(time)
-                        .withTimer(timer)
-                        .withMetadataImage(metadataImage)
-                        .withConsumerGroupHeartbeatInterval(5000)
-                        .withConsumerGroupSessionTimeout(45000)
-                        .withConsumerGroupMaxSize(consumerGroupMaxSize)
-                        .withConsumerGroupAssignors(consumerGroupAssignors)
-                        .withConsumerGroupMetadataRefreshIntervalMs(consumerGroupMetadataRefreshIntervalMs)
-                        .withClassicGroupMaxSize(classicGroupMaxSize)
-                        .withClassicGroupMinSessionTimeoutMs(classicGroupMinSessionTimeoutMs)
-                        .withClassicGroupMaxSessionTimeoutMs(classicGroupMaxSessionTimeoutMs)
-                        .withClassicGroupInitialRebalanceDelayMs(classicGroupInitialRebalanceDelayMs)
-                        .withClassicGroupNewMemberJoinTimeoutMs(classicGroupNewMemberJoinTimeoutMs)
-                        .withGroupCoordinatorMetricsShard(metrics)
-                        .build(),
-                    classicGroupInitialRebalanceDelayMs,
-                    classicGroupNewMemberJoinTimeoutMs
-                );
-
-                consumerGroupBuilders.forEach(builder -> {
-                    builder.build(metadataImage.topics()).forEach(context::replay);
-                });
-
-                context.commit();
-
-                return context;
-            }
-        }
-
-        final MockTime time;
-        final MockCoordinatorTimer<Void, Record> timer;
-        final SnapshotRegistry snapshotRegistry;
-        final GroupCoordinatorMetricsShard metrics;
-        final GroupMetadataManager groupMetadataManager;
-        final int classicGroupInitialRebalanceDelayMs;
-        final int classicGroupNewMemberJoinTimeoutMs;
-
-        long lastCommittedOffset = 0L;
-        long lastWrittenOffset = 0L;
-
-        public GroupMetadataManagerTestContext(
-            MockTime time,
-            MockCoordinatorTimer<Void, Record> timer,
-            SnapshotRegistry snapshotRegistry,
-            GroupCoordinatorMetricsShard metrics,
-            GroupMetadataManager groupMetadataManager,
-            int classicGroupInitialRebalanceDelayMs,
-            int classicGroupNewMemberJoinTimeoutMs
-        ) {
-            this.time = time;
-            this.timer = timer;
-            this.snapshotRegistry = snapshotRegistry;
-            this.metrics = metrics;
-            this.groupMetadataManager = groupMetadataManager;
-            this.classicGroupInitialRebalanceDelayMs = classicGroupInitialRebalanceDelayMs;
-            this.classicGroupNewMemberJoinTimeoutMs = classicGroupNewMemberJoinTimeoutMs;
-            snapshotRegistry.getOrCreateSnapshot(lastWrittenOffset);
-        }
-
-        public void commit() {
-            long lastCommittedOffset = this.lastCommittedOffset;
-            this.lastCommittedOffset = lastWrittenOffset;
-            snapshotRegistry.deleteSnapshotsUpTo(lastCommittedOffset);
-        }
-
-        public void rollback() {
-            lastWrittenOffset = lastCommittedOffset;
-            snapshotRegistry.revertToSnapshot(lastCommittedOffset);
-        }
-
-        public ConsumerGroup.ConsumerGroupState consumerGroupState(
-            String groupId
-        ) {
-            return groupMetadataManager
-                .getOrMaybeCreateConsumerGroup(groupId, false)
-                .state();
-        }
-
-        public ConsumerGroupMember.MemberState consumerGroupMemberState(
-            String groupId,
-            String memberId
-        ) {
-            return groupMetadataManager
-                .getOrMaybeCreateConsumerGroup(groupId, false)
-                .getOrMaybeCreateMember(memberId, false)
-                .state();
-        }
-
-        public CoordinatorResult<ConsumerGroupHeartbeatResponseData, Record> consumerGroupHeartbeat(
-            ConsumerGroupHeartbeatRequestData request
-        ) {
-            RequestContext context = new RequestContext(
-                new RequestHeader(
-                    ApiKeys.CONSUMER_GROUP_HEARTBEAT,
-                    ApiKeys.CONSUMER_GROUP_HEARTBEAT.latestVersion(),
-                    "client",
-                    0
-                ),
-                "1",
-                InetAddress.getLoopbackAddress(),
-                KafkaPrincipal.ANONYMOUS,
-                ListenerName.forSecurityProtocol(SecurityProtocol.PLAINTEXT),
-                SecurityProtocol.PLAINTEXT,
-                ClientInformation.EMPTY,
-                false
-            );
-
-            CoordinatorResult<ConsumerGroupHeartbeatResponseData, Record> result = groupMetadataManager.consumerGroupHeartbeat(
-                context,
-                request
-            );
-
-            result.records().forEach(this::replay);
-            return result;
-        }
-
-        public List<ExpiredTimeout<Void, Record>> sleep(long ms) {
-            time.sleep(ms);
-            List<ExpiredTimeout<Void, Record>> timeouts = timer.poll();
-            timeouts.forEach(timeout -> {
-                if (timeout.result.replayRecords()) {
-                    timeout.result.records().forEach(this::replay);
-                }
-            });
-            return timeouts;
-        }
-
-        public ScheduledTimeout<Void, Record> assertSessionTimeout(
-            String groupId,
-            String memberId,
-            long delayMs
-        ) {
-            ScheduledTimeout<Void, Record> timeout =
-                timer.timeout(consumerGroupSessionTimeoutKey(groupId, memberId));
-            assertNotNull(timeout);
-            assertEquals(time.milliseconds() + delayMs, timeout.deadlineMs);
-            return timeout;
-        }
-
-        public void assertNoSessionTimeout(
-            String groupId,
-            String memberId
-        ) {
-            ScheduledTimeout<Void, Record> timeout =
-                timer.timeout(consumerGroupSessionTimeoutKey(groupId, memberId));
-            assertNull(timeout);
-        }
-
-        public ScheduledTimeout<Void, Record> assertRevocationTimeout(
-            String groupId,
-            String memberId,
-            long delayMs
-        ) {
-            ScheduledTimeout<Void, Record> timeout =
-                timer.timeout(consumerGroupRevocationTimeoutKey(groupId, memberId));
-            assertNotNull(timeout);
-            assertEquals(time.milliseconds() + delayMs, timeout.deadlineMs);
-            return timeout;
-        }
-
-        public void assertNoRevocationTimeout(
-            String groupId,
-            String memberId
-        ) {
-            ScheduledTimeout<Void, Record> timeout =
-                timer.timeout(consumerGroupRevocationTimeoutKey(groupId, memberId));
-            assertNull(timeout);
-        }
-
-        ClassicGroup createClassicGroup(String groupId) {
-            return groupMetadataManager.getOrMaybeCreateClassicGroup(groupId, true);
-        }
-
-        public JoinResult sendClassicGroupJoin(
-            JoinGroupRequestData request
-        ) {
-            return sendClassicGroupJoin(request, false);
-        }
-
-        public JoinResult sendClassicGroupJoin(
-            JoinGroupRequestData request,
-            boolean requireKnownMemberId
-        ) {
-            return sendClassicGroupJoin(request, requireKnownMemberId, false);
-        }
-
-        public JoinResult sendClassicGroupJoin(
-            JoinGroupRequestData request,
-            boolean requireKnownMemberId,
-            boolean supportSkippingAssignment
-        ) {
-            // requireKnownMemberId is true: version >= 4 (See JoinGroupRequest#requiresKnownMemberId())
-            // supportSkippingAssignment is true: version >= 9 (See JoinGroupRequest#supportsSkippingAssignment())
-            short joinGroupVersion = 3;
-
-            if (requireKnownMemberId) {
-                joinGroupVersion = 4;
-                if (supportSkippingAssignment) {
-                    joinGroupVersion = ApiKeys.JOIN_GROUP.latestVersion();
-                }
-            }
-
-            RequestContext context = new RequestContext(
-                new RequestHeader(
-                    ApiKeys.JOIN_GROUP,
-                    joinGroupVersion,
-                    "client",
-                    0
-                ),
-                "1",
-                InetAddress.getLoopbackAddress(),
-                KafkaPrincipal.ANONYMOUS,
-                ListenerName.forSecurityProtocol(SecurityProtocol.PLAINTEXT),
-                SecurityProtocol.PLAINTEXT,
-                ClientInformation.EMPTY,
-                false
-            );
-
-            CompletableFuture<JoinGroupResponseData> responseFuture = new CompletableFuture<>();
-            CoordinatorResult<Void, Record> coordinatorResult = groupMetadataManager.classicGroupJoin(
-                context,
-                request,
-                responseFuture
-            );
-
-            return new JoinResult(responseFuture, coordinatorResult);
-        }
-
-        public JoinGroupResponseData joinClassicGroupAsDynamicMemberAndCompleteRebalance(
-            String groupId
-        ) throws Exception {
-            ClassicGroup group = createClassicGroup(groupId);
-
-            JoinGroupResponseData leaderJoinResponse =
-                joinClassicGroupAsDynamicMemberAndCompleteJoin(new JoinGroupRequestBuilder()
-                    .withGroupId("group-id")
-                    .withMemberId(UNKNOWN_MEMBER_ID)
-                    .withDefaultProtocolTypeAndProtocols()
-                    .withRebalanceTimeoutMs(10000)
-                    .withSessionTimeoutMs(5000)
-                    .build());
-
-            assertEquals(1, leaderJoinResponse.generationId());
-            assertTrue(group.isInState(COMPLETING_REBALANCE));
-
-            SyncResult syncResult = sendClassicGroupSync(new SyncGroupRequestBuilder()
-                .withGroupId("group-id")
-                .withMemberId(leaderJoinResponse.memberId())
-                .withGenerationId(leaderJoinResponse.generationId())
-                .build());
-
-            assertEquals(
-                Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
-                syncResult.records
-            );
-            // Simulate a successful write to the log.
-            syncResult.appendFuture.complete(null);
-
-            assertTrue(syncResult.syncFuture.isDone());
-            assertEquals(Errors.NONE.code(), syncResult.syncFuture.get().errorCode());
-            assertTrue(group.isInState(STABLE));
-
-            return leaderJoinResponse;
-        }
-
-        public JoinGroupResponseData joinClassicGroupAsDynamicMemberAndCompleteJoin(
-            JoinGroupRequestData request
-        ) throws ExecutionException, InterruptedException {
-            boolean requireKnownMemberId = true;
-            String newMemberId = request.memberId();
-
-            if (request.memberId().equals(UNKNOWN_MEMBER_ID)) {
-                // Since member id is required, we need another round to get the successful join group result.
-                JoinResult firstJoinResult = sendClassicGroupJoin(
-                    request,
-                    requireKnownMemberId
-                );
-                assertTrue(firstJoinResult.records.isEmpty());
-                assertTrue(firstJoinResult.joinFuture.isDone());
-                assertEquals(Errors.MEMBER_ID_REQUIRED.code(), firstJoinResult.joinFuture.get().errorCode());
-                newMemberId = firstJoinResult.joinFuture.get().memberId();
-            }
-
-            // Second round
-            JoinGroupRequestData secondRequest = new JoinGroupRequestData()
-                .setGroupId(request.groupId())
-                .setMemberId(newMemberId)
-                .setProtocolType(request.protocolType())
-                .setProtocols(request.protocols())
-                .setSessionTimeoutMs(request.sessionTimeoutMs())
-                .setRebalanceTimeoutMs(request.rebalanceTimeoutMs())
-                .setReason(request.reason());
-
-            JoinResult secondJoinResult = sendClassicGroupJoin(
-                secondRequest,
-                requireKnownMemberId
-            );
-
-            assertTrue(secondJoinResult.records.isEmpty());
-            List<ExpiredTimeout<Void, Record>> timeouts = sleep(classicGroupInitialRebalanceDelayMs);
-            assertEquals(1, timeouts.size());
-            assertTrue(secondJoinResult.joinFuture.isDone());
-            assertEquals(Errors.NONE.code(), secondJoinResult.joinFuture.get().errorCode());
-
-            return secondJoinResult.joinFuture.get();
-        }
-
-        public JoinGroupResponseData joinClassicGroupAndCompleteJoin(
-            JoinGroupRequestData request,
-            boolean requireKnownMemberId,
-            boolean supportSkippingAssignment
-        ) throws ExecutionException, InterruptedException {
-            return joinClassicGroupAndCompleteJoin(
-                request,
-                requireKnownMemberId,
-                supportSkippingAssignment,
-                classicGroupInitialRebalanceDelayMs
-            );
-        }
-
-        public JoinGroupResponseData joinClassicGroupAndCompleteJoin(
-            JoinGroupRequestData request,
-            boolean requireKnownMemberId,
-            boolean supportSkippingAssignment,
-            int advanceClockMs
-        ) throws ExecutionException, InterruptedException {
-            if (requireKnownMemberId && request.groupInstanceId().isEmpty()) {
-                return joinClassicGroupAsDynamicMemberAndCompleteJoin(request);
-            }
-
-            try {
-                JoinResult joinResult = sendClassicGroupJoin(
-                    request,
-                    requireKnownMemberId,
-                    supportSkippingAssignment
-                );
-
-                sleep(advanceClockMs);
-                assertTrue(joinResult.joinFuture.isDone());
-                assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
-                return joinResult.joinFuture.get();
-            } catch (Exception e) {
-                fail("Failed to due: " + e.getMessage());
-            }
-            return null;
-        }
-
-        public SyncResult sendClassicGroupSync(SyncGroupRequestData request) {
-            RequestContext context = new RequestContext(
-                new RequestHeader(
-                    ApiKeys.SYNC_GROUP,
-                    ApiKeys.SYNC_GROUP.latestVersion(),
-                    "client",
-                    0
-                ),
-                "1",
-                InetAddress.getLoopbackAddress(),
-                KafkaPrincipal.ANONYMOUS,
-                ListenerName.forSecurityProtocol(SecurityProtocol.PLAINTEXT),
-                SecurityProtocol.PLAINTEXT,
-                ClientInformation.EMPTY,
-                false
-            );
-
-            CompletableFuture<SyncGroupResponseData> responseFuture = new CompletableFuture<>();
-
-            CoordinatorResult<Void, Record> coordinatorResult = groupMetadataManager.classicGroupSync(
-                context,
-                request,
-                responseFuture
-            );
-
-            return new SyncResult(responseFuture, coordinatorResult);
-        }
-
-        public RebalanceResult staticMembersJoinAndRebalance(
-            String groupId,
-            String leaderInstanceId,
-            String followerInstanceId
-        ) throws Exception {
-            return staticMembersJoinAndRebalance(
-                groupId,
-                leaderInstanceId,
-                followerInstanceId,
-                10000,
-                5000
-            );
-        }
-
-        public RebalanceResult staticMembersJoinAndRebalance(
-            String groupId,
-            String leaderInstanceId,
-            String followerInstanceId,
-            int rebalanceTimeoutMs,
-            int sessionTimeoutMs
-        ) throws Exception {
-            ClassicGroup group = createClassicGroup("group-id");
-
-            JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
-                .withGroupId(groupId)
-                .withGroupInstanceId(leaderInstanceId)
-                .withMemberId(UNKNOWN_MEMBER_ID)
-                .withProtocolType("consumer")
-                .withProtocolSuperset()
-                .withRebalanceTimeoutMs(rebalanceTimeoutMs)
-                .withSessionTimeoutMs(sessionTimeoutMs)
-                .build();
-
-            JoinResult leaderJoinResult = sendClassicGroupJoin(joinRequest);
-            JoinResult followerJoinResult = sendClassicGroupJoin(joinRequest.setGroupInstanceId(followerInstanceId));
-
-            assertTrue(leaderJoinResult.records.isEmpty());
-            assertTrue(followerJoinResult.records.isEmpty());
-            assertFalse(leaderJoinResult.joinFuture.isDone());
-            assertFalse(followerJoinResult.joinFuture.isDone());
-
-            // The goal for two timer advance is to let first group initial join complete and set newMemberAdded flag to false. Next advance is
-            // to trigger the rebalance as needed for follower delayed join. One large time advance won't help because we could only populate one
-            // delayed join from purgatory and the new delayed op is created at that time and never be triggered.
-            assertNoOrEmptyResult(sleep(classicGroupInitialRebalanceDelayMs));
-            assertNoOrEmptyResult(sleep(classicGroupInitialRebalanceDelayMs));
-
-            assertTrue(leaderJoinResult.joinFuture.isDone());
-            assertTrue(followerJoinResult.joinFuture.isDone());
-            assertEquals(Errors.NONE.code(), leaderJoinResult.joinFuture.get().errorCode());
-            assertEquals(Errors.NONE.code(), followerJoinResult.joinFuture.get().errorCode());
-            assertEquals(1, leaderJoinResult.joinFuture.get().generationId());
-            assertEquals(1, followerJoinResult.joinFuture.get().generationId());
-            assertEquals(2, group.size());
-            assertEquals(1, group.generationId());
-            assertTrue(group.isInState(COMPLETING_REBALANCE));
-
-            String leaderId = leaderJoinResult.joinFuture.get().memberId();
-            String followerId = followerJoinResult.joinFuture.get().memberId();
-            List<SyncGroupRequestAssignment> assignment = new ArrayList<>();
-            assignment.add(new SyncGroupRequestAssignment().setMemberId(leaderId)
-                .setAssignment(new byte[]{1}));
-            assignment.add(new SyncGroupRequestAssignment().setMemberId(followerId)
-                .setAssignment(new byte[]{2}));
-
-            SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
-                .withGroupId(groupId)
-                .withGroupInstanceId(leaderInstanceId)
-                .withMemberId(leaderId)
-                .withGenerationId(1)
-                .withAssignment(assignment)
-                .build();
-
-            SyncResult leaderSyncResult = sendClassicGroupSync(syncRequest);
-
-            // The generated record should contain the new assignment.
-            Map<String, byte[]> groupAssignment = assignment.stream().collect(Collectors.toMap(
-                SyncGroupRequestAssignment::memberId, SyncGroupRequestAssignment::assignment
-            ));
-            assertEquals(
-                Collections.singletonList(
-                    RecordHelpers.newGroupMetadataRecord(group, groupAssignment, MetadataVersion.latestTesting())),
-                leaderSyncResult.records
-            );
-
-            // Simulate a successful write to the log.
-            leaderSyncResult.appendFuture.complete(null);
-
-            assertTrue(leaderSyncResult.syncFuture.isDone());
-            assertEquals(Errors.NONE.code(), leaderSyncResult.syncFuture.get().errorCode());
-            assertTrue(group.isInState(STABLE));
-
-            SyncResult followerSyncResult = sendClassicGroupSync(
-                syncRequest.setGroupInstanceId(followerInstanceId)
-                    .setMemberId(followerId)
-                    .setAssignments(Collections.emptyList())
-            );
-
-            assertTrue(followerSyncResult.records.isEmpty());
-            assertTrue(followerSyncResult.syncFuture.isDone());
-            assertEquals(Errors.NONE.code(), followerSyncResult.syncFuture.get().errorCode());
-            assertTrue(group.isInState(STABLE));
-
-            assertEquals(2, group.size());
-            assertEquals(1, group.generationId());
-
-            return new RebalanceResult(
-                1,
-                leaderId,
-                leaderSyncResult.syncFuture.get().assignment(),
-                followerId,
-                followerSyncResult.syncFuture.get().assignment()
-            );
-        }
-
-        public PendingMemberGroupResult setupGroupWithPendingMember(ClassicGroup group) throws Exception {
-            // Add the first member
-            JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
-                .withGroupId("group-id")
-                .withMemberId(UNKNOWN_MEMBER_ID)
-                .withDefaultProtocolTypeAndProtocols()
-                .withRebalanceTimeoutMs(10000)
-                .withSessionTimeoutMs(5000)
-                .build();
-
-            JoinGroupResponseData leaderJoinResponse =
-                joinClassicGroupAsDynamicMemberAndCompleteJoin(joinRequest);
-
-            List<SyncGroupRequestAssignment> assignment = new ArrayList<>();
-            assignment.add(new SyncGroupRequestAssignment().setMemberId(leaderJoinResponse.memberId()));
-            SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
-                .withGroupId("group-id")
-                .withMemberId(leaderJoinResponse.memberId())
-                .withGenerationId(leaderJoinResponse.generationId())
-                .withAssignment(assignment)
-                .build();
-
-            SyncResult syncResult = sendClassicGroupSync(syncRequest);
-
-            // Now the group is stable, with the one member that joined above
-            assertEquals(
-                Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
-                syncResult.records
-            );
-            // Simulate a successful write to log.
-            syncResult.appendFuture.complete(null);
-
-            assertTrue(syncResult.syncFuture.isDone());
-            assertEquals(Errors.NONE.code(), syncResult.syncFuture.get().errorCode());
-
-            // Start the join for the second member
-            JoinResult followerJoinResult = sendClassicGroupJoin(
-                joinRequest.setMemberId(UNKNOWN_MEMBER_ID)
-            );
-
-            assertTrue(followerJoinResult.records.isEmpty());
-            assertFalse(followerJoinResult.joinFuture.isDone());
-
-            JoinResult leaderJoinResult = sendClassicGroupJoin(
-                joinRequest.setMemberId(leaderJoinResponse.memberId())
-            );
-
-            assertTrue(leaderJoinResult.records.isEmpty());
-            assertTrue(group.isInState(COMPLETING_REBALANCE));
-            assertTrue(leaderJoinResult.joinFuture.isDone());
-            assertTrue(followerJoinResult.joinFuture.isDone());
-            assertEquals(Errors.NONE.code(), leaderJoinResult.joinFuture.get().errorCode());
-            assertEquals(Errors.NONE.code(), followerJoinResult.joinFuture.get().errorCode());
-            assertEquals(leaderJoinResult.joinFuture.get().generationId(), followerJoinResult.joinFuture.get().generationId());
-            assertEquals(leaderJoinResponse.memberId(), leaderJoinResult.joinFuture.get().leader());
-            assertEquals(leaderJoinResponse.memberId(), followerJoinResult.joinFuture.get().leader());
-
-            int nextGenerationId = leaderJoinResult.joinFuture.get().generationId();
-            String followerId = followerJoinResult.joinFuture.get().memberId();
-
-            // Stabilize the group
-            syncResult = sendClassicGroupSync(syncRequest.setGenerationId(nextGenerationId));
-
-            assertEquals(
-                Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
-                syncResult.records
-            );
-            // Simulate a successful write to log.
-            syncResult.appendFuture.complete(null);
-
-            assertTrue(syncResult.syncFuture.isDone());
-            assertEquals(Errors.NONE.code(), syncResult.syncFuture.get().errorCode());
-            assertTrue(group.isInState(STABLE));
-
-            // Re-join an existing member, to transition the group to PreparingRebalance state.
-            leaderJoinResult = sendClassicGroupJoin(
-                joinRequest.setMemberId(leaderJoinResponse.memberId()));
-
-            assertTrue(leaderJoinResult.records.isEmpty());
-            assertFalse(leaderJoinResult.joinFuture.isDone());
-            assertTrue(group.isInState(PREPARING_REBALANCE));
-
-            // Create a pending member in the group
-            JoinResult pendingMemberJoinResult = sendClassicGroupJoin(
-                joinRequest
-                    .setMemberId(UNKNOWN_MEMBER_ID)
-                    .setSessionTimeoutMs(2500),
-                true
-            );
-
-            assertTrue(pendingMemberJoinResult.records.isEmpty());
-            assertTrue(pendingMemberJoinResult.joinFuture.isDone());
-            assertEquals(Errors.MEMBER_ID_REQUIRED.code(), pendingMemberJoinResult.joinFuture.get().errorCode());
-            assertEquals(1, group.numPendingJoinMembers());
-
-            // Re-join the second existing member
-            followerJoinResult = sendClassicGroupJoin(
-                joinRequest.setMemberId(followerId).setSessionTimeoutMs(5000)
-            );
-
-            assertTrue(followerJoinResult.records.isEmpty());
-            assertFalse(followerJoinResult.joinFuture.isDone());
-            assertTrue(group.isInState(PREPARING_REBALANCE));
-            assertEquals(2, group.size());
-            assertEquals(1, group.numPendingJoinMembers());
-
-            return new PendingMemberGroupResult(
-                leaderJoinResponse.memberId(),
-                followerId,
-                pendingMemberJoinResult.joinFuture.get()
-            );
-        }
-
-        public void verifySessionExpiration(ClassicGroup group, int timeoutMs) {
-            Set<String> expectedHeartbeatKeys = group.allMembers().stream()
-                                                     .map(member -> classicGroupHeartbeatKey(group.groupId(), member.memberId())).collect(Collectors.toSet());
-
-            // Member should be removed as session expires.
-            List<ExpiredTimeout<Void, Record>> timeouts = sleep(timeoutMs);
-            List<Record> expectedRecords = Collections.singletonList(newGroupMetadataRecord(
-                group.groupId(),
-                new GroupMetadataValue()
-                    .setMembers(Collections.emptyList())
-                    .setGeneration(group.generationId())
-                    .setLeader(null)
-                    .setProtocolType("consumer")
-                    .setProtocol(null)
-                    .setCurrentStateTimestamp(time.milliseconds()),
-                MetadataVersion.latestTesting()));
-
-
-            Set<String> heartbeatKeys = timeouts.stream().map(timeout -> timeout.key).collect(Collectors.toSet());
-            assertEquals(expectedHeartbeatKeys, heartbeatKeys);
-
-            // Only the last member leaving the group should result in the empty group metadata record.
-            int timeoutsSize = timeouts.size();
-            assertEquals(expectedRecords, timeouts.get(timeoutsSize - 1).result.records());
-            assertNoOrEmptyResult(timeouts.subList(0, timeoutsSize - 1));
-            assertTrue(group.isInState(EMPTY));
-            assertEquals(0, group.size());
-        }
-
-        public HeartbeatResponseData sendClassicGroupHeartbeat(
-            HeartbeatRequestData request
-        ) {
-            RequestContext context = new RequestContext(
-                new RequestHeader(
-                    ApiKeys.HEARTBEAT,
-                    ApiKeys.HEARTBEAT.latestVersion(),
-                    "client",
-                    0
-                ),
-                "1",
-                InetAddress.getLoopbackAddress(),
-                KafkaPrincipal.ANONYMOUS,
-                ListenerName.forSecurityProtocol(SecurityProtocol.PLAINTEXT),
-                SecurityProtocol.PLAINTEXT,
-                ClientInformation.EMPTY,
-                false
-            );
-
-            return groupMetadataManager.classicGroupHeartbeat(
-                context,
-                request
-            );
-        }
-
-        public List<ListGroupsResponseData.ListedGroup> sendListGroups(List<String> statesFilter, List<String> typesFilter) {
-            Set<String> statesFilterSet = statesFilter.stream().collect(Collectors.toSet());
-            Set<String> typesFilterSet = typesFilter.stream().collect(Collectors.toSet());
-            return groupMetadataManager.listGroups(statesFilterSet, typesFilterSet, lastCommittedOffset);
-        }
-
-        public List<ConsumerGroupDescribeResponseData.DescribedGroup> sendConsumerGroupDescribe(List<String> groupIds) {
-            return groupMetadataManager.consumerGroupDescribe(groupIds, lastCommittedOffset);
-        }
-
-        public List<DescribeGroupsResponseData.DescribedGroup> describeGroups(List<String> groupIds) {
-            return groupMetadataManager.describeGroups(groupIds, lastCommittedOffset);
-        }
-
-        public void verifyHeartbeat(
-            String groupId,
-            JoinGroupResponseData joinResponse,
-            Errors expectedError
-        ) {
-            HeartbeatRequestData request = new HeartbeatRequestData()
-                .setGroupId(groupId)
-                .setMemberId(joinResponse.memberId())
-                .setGenerationId(joinResponse.generationId());
-
-            if (expectedError == Errors.UNKNOWN_MEMBER_ID) {
-                assertThrows(UnknownMemberIdException.class, () -> sendClassicGroupHeartbeat(request));
-            } else {
-                HeartbeatResponseData response = sendClassicGroupHeartbeat(request);
-                assertEquals(expectedError.code(), response.errorCode());
-            }
-        }
-
-        public List<JoinGroupResponseData> joinWithNMembers(
-            String groupId,
-            int numMembers,
-            int rebalanceTimeoutMs,
-            int sessionTimeoutMs
-        ) {
-            ClassicGroup group = createClassicGroup(groupId);
-            boolean requireKnownMemberId = true;
-
-            // First join requests
-            JoinGroupRequestData request = new JoinGroupRequestBuilder()
-                .withGroupId(groupId)
-                .withMemberId(UNKNOWN_MEMBER_ID)
-                .withDefaultProtocolTypeAndProtocols()
-                .withRebalanceTimeoutMs(rebalanceTimeoutMs)
-                .withSessionTimeoutMs(sessionTimeoutMs)
-                .build();
-
-            List<String> memberIds = IntStream.range(0, numMembers).mapToObj(i -> {
-                JoinResult joinResult = sendClassicGroupJoin(request, requireKnownMemberId);
-
-                assertTrue(joinResult.records.isEmpty());
-                assertTrue(joinResult.joinFuture.isDone());
-
-                try {
-                    return joinResult.joinFuture.get().memberId();
-                } catch (Exception e) {
-                    fail("Unexpected exception: " + e.getMessage());
-                }
-                return null;
-            }).collect(Collectors.toList());
-
-            // Second join requests
-            List<CompletableFuture<JoinGroupResponseData>> secondJoinFutures = IntStream.range(0, numMembers).mapToObj(i -> {
-                JoinResult joinResult = sendClassicGroupJoin(request.setMemberId(memberIds.get(i)), requireKnownMemberId);
-
-                assertTrue(joinResult.records.isEmpty());
-                assertFalse(joinResult.joinFuture.isDone());
-
-                return joinResult.joinFuture;
-            }).collect(Collectors.toList());
-
-            // Advance clock by initial rebalance delay.
-            assertNoOrEmptyResult(sleep(classicGroupInitialRebalanceDelayMs));
-            secondJoinFutures.forEach(future -> assertFalse(future.isDone()));
-            // Advance clock by rebalance timeout to complete join phase.
-            assertNoOrEmptyResult(sleep(rebalanceTimeoutMs));
-
-            List<JoinGroupResponseData> joinResponses = secondJoinFutures.stream().map(future -> {
-                assertTrue(future.isDone());
-                try {
-                    assertEquals(Errors.NONE.code(), future.get().errorCode());
-                    return future.get();
-                } catch (Exception e) {
-                    fail("Unexpected exception: " + e.getMessage());
-                }
-                return null;
-            }).collect(Collectors.toList());
-
-            assertEquals(numMembers, group.size());
-            assertTrue(group.isInState(COMPLETING_REBALANCE));
-
-            return joinResponses;
-        }
-
-        public CoordinatorResult<LeaveGroupResponseData, Record> sendClassicGroupLeave(
-            LeaveGroupRequestData request
-        ) {
-            RequestContext context = new RequestContext(
-                new RequestHeader(
-                    ApiKeys.LEAVE_GROUP,
-                    ApiKeys.LEAVE_GROUP.latestVersion(),
-                    "client",
-                    0
-                ),
-                "1",
-                InetAddress.getLoopbackAddress(),
-                KafkaPrincipal.ANONYMOUS,
-                ListenerName.forSecurityProtocol(SecurityProtocol.PLAINTEXT),
-                SecurityProtocol.PLAINTEXT,
-                ClientInformation.EMPTY,
-                false
-            );
-
-            return groupMetadataManager.classicGroupLeave(context, request);
-        }
-
-        private void verifyDescribeGroupsReturnsDeadGroup(String groupId) {
-            List<DescribeGroupsResponseData.DescribedGroup> describedGroups =
-                describeGroups(Collections.singletonList(groupId));
-
-            assertEquals(
-                Collections.singletonList(new DescribeGroupsResponseData.DescribedGroup()
-                    .setGroupId("group-id")
-                    .setGroupState(DEAD.toString())
-                ),
-                describedGroups
-            );
-        }
-
-        private ApiMessage messageOrNull(ApiMessageAndVersion apiMessageAndVersion) {
-            if (apiMessageAndVersion == null) {
-                return null;
-            } else {
-                return apiMessageAndVersion.message();
-            }
-        }
-
-        private void replay(
-            Record record
-        ) {
-            ApiMessageAndVersion key = record.key();
-            ApiMessageAndVersion value = record.value();
-
-            if (key == null) {
-                throw new IllegalStateException("Received a null key in " + record);
-            }
-
-            switch (key.version()) {
-                case GroupMetadataKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (GroupMetadataKey) key.message(),
-                        (GroupMetadataValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupMemberMetadataKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupMemberMetadataKey) key.message(),
-                        (ConsumerGroupMemberMetadataValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupMetadataKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupMetadataKey) key.message(),
-                        (ConsumerGroupMetadataValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupPartitionMetadataKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupPartitionMetadataKey) key.message(),
-                        (ConsumerGroupPartitionMetadataValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupTargetAssignmentMemberKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupTargetAssignmentMemberKey) key.message(),
-                        (ConsumerGroupTargetAssignmentMemberValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupTargetAssignmentMetadataKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupTargetAssignmentMetadataKey) key.message(),
-                        (ConsumerGroupTargetAssignmentMetadataValue) messageOrNull(value)
-                    );
-                    break;
-
-                case ConsumerGroupCurrentMemberAssignmentKey.HIGHEST_SUPPORTED_VERSION:
-                    groupMetadataManager.replay(
-                        (ConsumerGroupCurrentMemberAssignmentKey) key.message(),
-                        (ConsumerGroupCurrentMemberAssignmentValue) messageOrNull(value)
-                    );
-                    break;
-
-                default:
-                    throw new IllegalStateException("Received an unknown record type " + key.version()
-                        + " in " + record);
-            }
-
-            lastWrittenOffset++;
-            snapshotRegistry.getOrCreateSnapshot(lastWrittenOffset);
-        }
-    }
 
     @Test
     public void testConsumerHeartbeatRequestValidation() {
@@ -1563,7 +440,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId)
                 .setMemberEpoch(1)
@@ -1609,7 +486,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
     }
 
     @Test
@@ -1662,7 +539,7 @@ public class GroupMetadataManagerTest {
                 .setMemberEpoch(10)
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar")));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId)
                 .setMemberEpoch(11)
@@ -1709,7 +586,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
     }
 
     @Test
@@ -1799,7 +676,7 @@ public class GroupMetadataManagerTest {
                 .setServerAssignor("range")
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -1847,9 +724,9 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember3)
         );
 
-        assertRecordsEquals(expectedRecords.subList(0, 3), result.records().subList(0, 3));
-        assertUnorderedListEquals(expectedRecords.subList(3, 6), result.records().subList(3, 6));
-        assertRecordsEquals(expectedRecords.subList(6, 8), result.records().subList(6, 8));
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords.subList(0, 3), result.records().subList(0, 3));
+        GroupMetadataManagerTestContext.assertUnorderedListEquals(expectedRecords.subList(3, 6), result.records().subList(3, 6));
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords.subList(6, 8), result.records().subList(6, 8));
     }
 
     @Test
@@ -1922,7 +799,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(LEAVE_GROUP_MEMBER_EPOCH),
@@ -1943,7 +820,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newGroupEpochRecord(groupId, 11)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
     }
 
     @Test
@@ -2037,7 +914,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -2086,9 +963,9 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember3)
         );
 
-        assertRecordsEquals(expectedRecords.subList(0, 3), result.records().subList(0, 3));
-        assertUnorderedListEquals(expectedRecords.subList(3, 6), result.records().subList(3, 6));
-        assertRecordsEquals(expectedRecords.subList(6, 8), result.records().subList(6, 8));
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords.subList(0, 3), result.records().subList(0, 3));
+        GroupMetadataManagerTestContext.assertUnorderedListEquals(expectedRecords.subList(3, 6), result.records().subList(3, 6));
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords.subList(6, 8), result.records().subList(6, 8));
     }
 
     @Test
@@ -2186,7 +1063,7 @@ public class GroupMetadataManagerTest {
                 .setTopicPartitions(Collections.emptyList()));
 
         // Member epoch of the response would be set to -2.
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(-2),
@@ -2199,7 +1076,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         assertEquals(1, result.records().size());
-        assertRecordEquals(result.records().get(0), RecordHelpers.newCurrentAssignmentRecord(groupId, member2UpdatedEpoch));
+        GroupMetadataManagerTestContext.assertRecordEquals(result.records().get(0), RecordHelpers.newCurrentAssignmentRecord(groupId, member2UpdatedEpoch));
 
         // Member 2 rejoins the group with the same instance id.
         CoordinatorResult<ConsumerGroupHeartbeatResponseData, Record> rejoinResult = context.consumerGroupHeartbeat(
@@ -2213,7 +1090,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(member2RejoinId)
                 .setMemberEpoch(10)
@@ -2257,7 +1134,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedRejoinedMember)
         );
 
-        assertRecordsEquals(expectedRecordsAfterRejoin, rejoinResult.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecordsAfterRejoin, rejoinResult.records());
         // Verify that there are no timers.
         context.assertNoSessionTimeout(groupId, memberId2);
         context.assertNoRevocationTimeout(groupId, memberId2);
@@ -2336,7 +1213,7 @@ public class GroupMetadataManagerTest {
                 .setTopicPartitions(Collections.emptyList()));
 
         // member epoch of the response would be set to -2
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(LEAVE_GROUP_STATIC_MEMBER_EPOCH),
@@ -2349,7 +1226,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         assertEquals(1, result.records().size());
-        assertRecordEquals(result.records().get(0), RecordHelpers.newCurrentAssignmentRecord(groupId, member2UpdatedEpoch));
+        GroupMetadataManagerTestContext.assertRecordEquals(result.records().get(0), RecordHelpers.newCurrentAssignmentRecord(groupId, member2UpdatedEpoch));
     }
 
     @Test
@@ -2425,7 +1302,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Arrays.asList("foo", "bar"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(LEAVE_GROUP_MEMBER_EPOCH),
@@ -2446,7 +1323,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newGroupEpochRecord(groupId, 11)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
     }
 
     @Test
@@ -2871,7 +1748,7 @@ public class GroupMetadataManagerTest {
                 .setServerAssignor("range")
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -2882,7 +1759,7 @@ public class GroupMetadataManagerTest {
 
         // We only check the last record as the subscription/target assignment updates are
         // already covered by other tests.
-        assertRecordEquals(
+        GroupMetadataManagerTestContext.assertRecordEquals(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId3)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(0)
@@ -2905,7 +1782,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId1)
             .setMemberEpoch(10));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(10)
@@ -2922,7 +1799,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId1)
                 .setMemberEpoch(10)
                 .setPreviousMemberEpoch(9)
@@ -2948,7 +1825,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId2)
             .setMemberEpoch(10));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(10)
@@ -2965,7 +1842,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId2)
                 .setMemberEpoch(10)
                 .setPreviousMemberEpoch(9)
@@ -2991,7 +1868,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId3)
             .setMemberEpoch(11));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -3019,7 +1896,7 @@ public class GroupMetadataManagerTest {
                     .setPartitions(Arrays.asList(0))
             )));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(11)
@@ -3036,7 +1913,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId1)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(10)
@@ -3057,7 +1934,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId2)
             .setMemberEpoch(10));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(10)
@@ -3076,7 +1953,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId3)
             .setMemberEpoch(11));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -3089,7 +1966,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId3)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(11)
@@ -3112,7 +1989,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId3)
             .setMemberEpoch(11));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -3140,7 +2017,7 @@ public class GroupMetadataManagerTest {
                     .setPartitions(Arrays.asList(2))
             )));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(11)
@@ -3157,7 +2034,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId2)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(10)
@@ -3178,7 +2055,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId3)
             .setMemberEpoch(11));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(11)
@@ -3194,7 +2071,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId3)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(11)
@@ -3274,7 +2151,7 @@ public class GroupMetadataManagerTest {
                 .setServerAssignor("range")
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(11)
@@ -3283,7 +2160,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordEquals(
+        GroupMetadataManagerTestContext.assertRecordEquals(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId2)
                 .setMemberEpoch(11)
                 .setPreviousMemberEpoch(0)
@@ -3302,7 +2179,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId1)
             .setMemberEpoch(10));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(10)
@@ -3315,7 +2192,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId1)
                 .setMemberEpoch(10)
                 .setPreviousMemberEpoch(9)
@@ -3358,7 +2235,7 @@ public class GroupMetadataManagerTest {
                 .setServerAssignor("range")
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(12)
@@ -3367,7 +2244,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordEquals(
+        GroupMetadataManagerTestContext.assertRecordEquals(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId3)
                 .setMemberEpoch(12)
                 .setPreviousMemberEpoch(0)
@@ -3386,7 +2263,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId1)
             .setMemberEpoch(10));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(10)
@@ -3399,7 +2276,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
                 RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId1)
                     .setMemberEpoch(10)
                     .setPreviousMemberEpoch(9)
@@ -3420,7 +2297,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(memberId2)
             .setMemberEpoch(11));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(12)
@@ -3429,7 +2306,7 @@ public class GroupMetadataManagerTest {
             result.response()
         );
 
-        assertRecordsEquals(Collections.singletonList(
+        GroupMetadataManagerTestContext.assertRecordsEquals(Collections.singletonList(
             RecordHelpers.newCurrentAssignmentRecord(groupId, new ConsumerGroupMember.Builder(memberId2)
                 .setMemberEpoch(12)
                 .setPreviousMemberEpoch(11)
@@ -3662,7 +2539,7 @@ public class GroupMetadataManagerTest {
                 .setMemberEpoch(10));
 
         // The member gets partitions 3, 4 and 5 assigned.
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId)
                 .setMemberEpoch(11)
@@ -3702,7 +2579,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
 
         // Check next refresh time.
         assertFalse(consumerGroup.hasMetadataExpired(context.time.milliseconds()));
@@ -3791,7 +2668,7 @@ public class GroupMetadataManagerTest {
 
 
         // The member gets partitions 3, 4 and 5 assigned.
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId)
                 .setMemberEpoch(11)
@@ -3831,7 +2708,7 @@ public class GroupMetadataManagerTest {
             RecordHelpers.newCurrentAssignmentRecord(groupId, expectedMember)
         );
 
-        assertRecordsEquals(expectedRecords, result.records());
+        GroupMetadataManagerTestContext.assertRecordsEquals(expectedRecords, result.records());
 
         // Check next refresh time.
         assertFalse(consumerGroup.hasMetadataExpired(context.time.milliseconds()));
@@ -4304,7 +3181,7 @@ public class GroupMetadataManagerTest {
                     .setSubscribedTopicNames(Collections.singletonList("foo"))
                     .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(1)
@@ -4346,7 +3223,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Collections.singletonList("foo"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(2)
@@ -4370,7 +3247,7 @@ public class GroupMetadataManagerTest {
                 .setRebalanceTimeoutMs(12000)
                 .setSubscribedTopicNames(Collections.singletonList("foo")));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(1)
@@ -4418,7 +3295,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Collections.singletonList("foo"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId3)
                 .setMemberEpoch(3)
@@ -4442,7 +3319,7 @@ public class GroupMetadataManagerTest {
                 .setRebalanceTimeoutMs(90000)
                 .setSubscribedTopicNames(Collections.singletonList("foo")));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(1)
@@ -4475,7 +3352,7 @@ public class GroupMetadataManagerTest {
                     .setTopicId(fooTopicId)
                     .setPartitions(Collections.singletonList(0)))));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(3)
@@ -4536,7 +3413,7 @@ public class GroupMetadataManagerTest {
                     .setSubscribedTopicNames(Collections.singletonList("foo"))
                     .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(1)
@@ -4578,7 +3455,7 @@ public class GroupMetadataManagerTest {
                 .setSubscribedTopicNames(Collections.singletonList("foo"))
                 .setTopicPartitions(Collections.emptyList()));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId2)
                 .setMemberEpoch(2)
@@ -4600,7 +3477,7 @@ public class GroupMetadataManagerTest {
                 .setMemberId(memberId1)
                 .setMemberEpoch(1));
 
-        assertResponseEquals(
+        GroupMetadataManagerTestContext.assertResponseEquals(
             new ConsumerGroupHeartbeatResponseData()
                 .setMemberId(memberId1)
                 .setMemberEpoch(1)
@@ -4699,13 +3576,13 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true);
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.MEMBER_ID_REQUIRED.code(), joinResult.joinFuture.get().errorCode());
 
@@ -4722,14 +3599,14 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withGroupInstanceId("group-instance-id")
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true);
         assertFalse(joinResult.joinFuture.isDone());
 
         // Simulate a failed write to the log.
@@ -4778,7 +3655,7 @@ public class GroupMetadataManagerTest {
             ));
         });
 
-        Record groupMetadataRecord = newGroupMetadataRecord("group-id",
+        Record groupMetadataRecord = GroupMetadataManagerTestContext.newGroupMetadataRecord("group-id",
             new GroupMetadataValue()
                 .setMembers(members)
                 .setGeneration(1)
@@ -4842,7 +3719,7 @@ public class GroupMetadataManagerTest {
             );
         });
 
-        Record groupMetadataRecord = newGroupMetadataRecord("group-id",
+        Record groupMetadataRecord = GroupMetadataManagerTestContext.newGroupMetadataRecord("group-id",
             new GroupMetadataValue()
                 .setMembers(members)
                 .setGeneration(1)
@@ -4882,7 +3759,7 @@ public class GroupMetadataManagerTest {
             );
         });
 
-        Record groupMetadataRecord = newGroupMetadataRecord("group-id",
+        Record groupMetadataRecord = GroupMetadataManagerTestContext.newGroupMetadataRecord("group-id",
             new GroupMetadataValue()
                 .setMembers(members)
                 .setGeneration(1)
@@ -4911,7 +3788,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -4919,12 +3796,12 @@ public class GroupMetadataManagerTest {
             .build();
 
         IntStream.range(0, 10).forEach(i -> {
-            JoinResult joinResult = context.sendClassicGroupJoin(request);
+            GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
             assertFalse(joinResult.joinFuture.isDone());
             assertTrue(joinResult.records.isEmpty());
         });
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.GROUP_MAX_SIZE_REACHED.code(), joinResult.joinFuture.get().errorCode());
@@ -4941,7 +3818,7 @@ public class GroupMetadataManagerTest {
 
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -4949,12 +3826,12 @@ public class GroupMetadataManagerTest {
 
         // First round of join requests. Generate member ids. All requests will be accepted
         // as the group is still Empty.
-        List<JoinResult> firstRoundJoinResults = IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> firstRoundJoinResults = IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
             request,
             requiredKnownMemberId
         )).collect(Collectors.toList());
 
-        List<String> memberIds = verifyClassicGroupJoinResponses(firstRoundJoinResults, 0, Errors.MEMBER_ID_REQUIRED);
+        List<String> memberIds = GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(firstRoundJoinResults, 0, Errors.MEMBER_ID_REQUIRED);
         assertEquals(groupMaxSize + 1, memberIds.size());
         assertEquals(0, group.size());
         assertTrue(group.isInState(EMPTY));
@@ -4962,31 +3839,31 @@ public class GroupMetadataManagerTest {
 
         // Second round of join requests with the generated member ids.
         // One of them will fail, reaching group max size.
-        List<JoinResult> secondRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> secondRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
             request.setMemberId(memberId),
             requiredKnownMemberId
         )).collect(Collectors.toList());
 
         // Advance clock by group initial rebalance delay to complete first inital delayed join.
         // This will extend the initial rebalance as new members have joined.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
         // Advance clock by group initial rebalance delay to complete second inital delayed join.
         // Since there are no new members that joined since the previous delayed join,
         // the join group phase will complete.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
 
-        verifyClassicGroupJoinResponses(secondRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(secondRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
         assertEquals(groupMaxSize, group.size());
         assertEquals(0, group.numPendingJoinMembers());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
         // Members that were accepted can rejoin while others are rejected in CompletingRebalance state.
-        List<JoinResult> thirdRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> thirdRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
             request.setMemberId(memberId),
             requiredKnownMemberId
         )).collect(Collectors.toList());
 
-        verifyClassicGroupJoinResponses(thirdRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(thirdRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
     }
 
     @Test
@@ -5000,14 +3877,14 @@ public class GroupMetadataManagerTest {
 
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
         // First round of join requests. This will trigger a rebalance.
-        List<JoinResult> firstRoundJoinResults = IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> firstRoundJoinResults = IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
             request,
             requiredKnownMemberId
         )).collect(Collectors.toList());
@@ -5018,21 +3895,21 @@ public class GroupMetadataManagerTest {
 
         // Advance clock by group initial rebalance delay to complete first inital delayed join.
         // This will extend the initial rebalance as new members have joined.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
         // Advance clock by group initial rebalance delay to complete second inital delayed join.
         // Since there are no new members that joined since the previous delayed join,
         // we will complete the rebalance.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
 
-        List<String> memberIds = verifyClassicGroupJoinResponses(firstRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        List<String> memberIds = GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(firstRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
 
         // Members that were accepted can rejoin while others are rejected in CompletingRebalance state.
-        List<JoinResult> secondRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> secondRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
             request.setMemberId(memberId),
             requiredKnownMemberId
         )).collect(Collectors.toList());
 
-        verifyClassicGroupJoinResponses(secondRoundJoinResults, 10, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(secondRoundJoinResults, 10, Errors.GROUP_MAX_SIZE_REACHED);
         assertEquals(groupMaxSize, group.size());
         assertEquals(0, group.numAwaitingJoinResponse());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
@@ -5053,16 +3930,16 @@ public class GroupMetadataManagerTest {
 
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
         // First round of join requests. This will trigger a rebalance.
-        List<JoinResult> firstRoundJoinResults = groupInstanceIds.stream()
-            .map(instanceId -> context.sendClassicGroupJoin(request.setGroupInstanceId(instanceId)))
-            .collect(Collectors.toList());
+        List<GroupMetadataManagerTestContext.JoinResult> firstRoundJoinResults = groupInstanceIds.stream()
+                                                                                                 .map(instanceId -> context.sendClassicGroupJoin(request.setGroupInstanceId(instanceId)))
+                                                                                                 .collect(Collectors.toList());
 
         assertEquals(groupMaxSize, group.size());
         assertEquals(groupMaxSize, group.numAwaitingJoinResponse());
@@ -5070,23 +3947,23 @@ public class GroupMetadataManagerTest {
 
         // Advance clock by group initial rebalance delay to complete first inital delayed join.
         // This will extend the initial rebalance as new members have joined.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
         // Advance clock by group initial rebalance delay to complete second inital delayed join.
         // Since there are no new members that joined since the previous delayed join,
         // we will complete the rebalance.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
 
-        List<String> memberIds = verifyClassicGroupJoinResponses(firstRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        List<String> memberIds = GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(firstRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
 
         // Members which were accepted can rejoin, others are rejected, while
         // completing rebalance
-        List<JoinResult> secondRoundJoinResults =  IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> secondRoundJoinResults =  IntStream.range(0, groupMaxSize + 1).mapToObj(i -> context.sendClassicGroupJoin(
             request
                 .setMemberId(memberIds.get(i))
                 .setGroupInstanceId(groupInstanceIds.get(i))
         )).collect(Collectors.toList());
 
-        verifyClassicGroupJoinResponses(secondRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(secondRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
         assertEquals(groupMaxSize, group.size());
         assertEquals(0, group.numAwaitingJoinResponse());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
@@ -5103,28 +3980,28 @@ public class GroupMetadataManagerTest {
 
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
         // First round of join requests. Generate member ids.
-        List<JoinResult> firstRoundJoinResults =  IntStream.range(0, groupMaxSize + 1)
-            .mapToObj(__ -> context.sendClassicGroupJoin(request, requiredKnownMemberId))
-            .collect(Collectors.toList());
+        List<GroupMetadataManagerTestContext.JoinResult> firstRoundJoinResults =  IntStream.range(0, groupMaxSize + 1)
+                                                                                           .mapToObj(__ -> context.sendClassicGroupJoin(request, requiredKnownMemberId))
+                                                                                           .collect(Collectors.toList());
 
         assertEquals(0, group.size());
         assertEquals(groupMaxSize + 1, group.numPendingJoinMembers());
         assertTrue(group.isInState(EMPTY));
 
-        List<String> memberIds = verifyClassicGroupJoinResponses(firstRoundJoinResults, 0, Errors.MEMBER_ID_REQUIRED);
+        List<String> memberIds = GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(firstRoundJoinResults, 0, Errors.MEMBER_ID_REQUIRED);
         assertEquals(groupMaxSize + 1, memberIds.size());
 
         // Second round of join requests with the generated member ids.
         // One of them will fail, reaching group max size.
         memberIds.forEach(memberId -> {
-            JoinResult joinResult = context.sendClassicGroupJoin(
+            GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
                 request.setMemberId(memberId),
                 requiredKnownMemberId
             );
@@ -5136,20 +4013,20 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Members can rejoin while rebalancing
-        List<JoinResult> thirdRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
+        List<GroupMetadataManagerTestContext.JoinResult> thirdRoundJoinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
             request.setMemberId(memberId),
             requiredKnownMemberId
         )).collect(Collectors.toList());
 
         // Advance clock by group initial rebalance delay to complete first inital delayed join.
         // This will extend the initial rebalance as new members have joined.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
         // Advance clock by group initial rebalance delay to complete second inital delayed join.
         // Since there are no new members that joined since the previous delayed join,
         // we will complete the rebalance.
-        assertNoOrEmptyResult(context.sleep(50));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(50));
 
-        verifyClassicGroupJoinResponses(thirdRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(thirdRoundJoinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
         assertEquals(groupMaxSize, group.size());
         assertEquals(0, group.numAwaitingJoinResponse());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
@@ -5181,15 +4058,15 @@ public class GroupMetadataManagerTest {
                     10000,
                     5000,
                     "consumer",
-                    toProtocols("range")
+                    GroupMetadataManagerTestContext.toProtocols("range")
                 )
             );
         });
 
         context.groupMetadataManager.prepareRebalance(group, "test");
 
-        List<JoinResult> joinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        List<GroupMetadataManagerTestContext.JoinResult> joinResults = memberIds.stream().map(memberId -> context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(memberId)
                 .withDefaultProtocolTypeAndProtocols()
@@ -5202,9 +4079,9 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Advance clock by rebalance timeout to complete join phase.
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
 
-        verifyClassicGroupJoinResponses(joinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
+        GroupMetadataManagerTestContext.verifyClassicGroupJoinResponses(joinResults, groupMaxSize, Errors.GROUP_MAX_SIZE_REACHED);
 
         assertEquals(groupMaxSize, group.size());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
@@ -5223,13 +4100,13 @@ public class GroupMetadataManagerTest {
             .withClassicGroupMinSessionTimeoutMs(minSessionTimeout)
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withSessionTimeoutMs(minSessionTimeout - 1)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
         assertTrue(joinResult.joinFuture.isDone());
         assertTrue(joinResult.records.isEmpty());
         assertEquals(Errors.INVALID_SESSION_TIMEOUT.code(), joinResult.joinFuture.get().errorCode());
@@ -5242,13 +4119,13 @@ public class GroupMetadataManagerTest {
             .withClassicGroupMaxSessionTimeoutMs(maxSessionTimeout)
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withSessionTimeoutMs(maxSessionTimeout + 1)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5260,19 +4137,19 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId("member-id")
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.UNKNOWN_MEMBER_ID.code(), joinResult.joinFuture.get().errorCode());
 
         // Static member
-        request = new JoinGroupRequestBuilder()
+        request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId("member-id")
             .withGroupInstanceId("group-instance-id")
@@ -5291,7 +4168,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -5299,14 +4176,14 @@ public class GroupMetadataManagerTest {
 
         context.joinClassicGroupAsDynamicMemberAndCompleteJoin(request);
 
-        request = new JoinGroupRequestBuilder()
+        request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("connect")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5319,14 +4196,14 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.INCONSISTENT_GROUP_PROTOCOL.code(), joinResult.joinFuture.get().errorCode());
@@ -5344,14 +4221,14 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
             .withProtocols(new JoinGroupRequestProtocolCollection(0))
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.INCONSISTENT_GROUP_PROTOCOL.code(), joinResult.joinFuture.get().errorCode());
@@ -5369,7 +4246,7 @@ public class GroupMetadataManagerTest {
 
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -5386,7 +4263,7 @@ public class GroupMetadataManagerTest {
         assertEquals(0, group.allMembers().stream().filter(ClassicGroupMember::isNew).count());
 
         // Send second join group request for a new dynamic member.
-        JoinResult secondJoinResult = context.sendClassicGroupJoin(request
+        GroupMetadataManagerTestContext.JoinResult secondJoinResult = context.sendClassicGroupJoin(request
             .setSessionTimeoutMs(5000)
             .setRebalanceTimeoutMs(5000)
         );
@@ -5400,7 +4277,7 @@ public class GroupMetadataManagerTest {
         assertNotEquals(firstMemberId, newMember.memberId());
 
         // Advance clock by new member join timeout to expire the second member.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupNewMemberJoinTimeoutMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupNewMemberJoinTimeoutMs));
 
         assertTrue(secondJoinResult.joinFuture.isDone());
         JoinGroupResponseData secondResponse = secondJoinResult.joinFuture.get();
@@ -5417,24 +4294,24 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
 
-        JoinResult otherJoinResult = context.sendClassicGroupJoin(request.setProtocols(toProtocols("roundrobin")));
+        GroupMetadataManagerTestContext.JoinResult otherJoinResult = context.sendClassicGroupJoin(request.setProtocols(GroupMetadataManagerTestContext.toProtocols("roundrobin")));
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(otherJoinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
         assertEquals(Errors.INCONSISTENT_GROUP_PROTOCOL.code(), otherJoinResult.joinFuture.get().errorCode());
@@ -5446,13 +4323,13 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5471,12 +4348,12 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.INCONSISTENT_GROUP_PROTOCOL.code(), joinResult.joinFuture.get().errorCode());
 
         // Sending consistent protocol should be accepted
-        joinResult = context.sendClassicGroupJoin(request.setProtocols(toProtocols("range")), true);
+        joinResult = context.sendClassicGroupJoin(request.setProtocols(GroupMetadataManagerTestContext.toProtocols("range")), true);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
 
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
@@ -5488,7 +4365,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withGroupInstanceId("group-instance-id")
@@ -5504,7 +4381,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withGroupInstanceId("group-instance-id")
@@ -5516,7 +4393,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData response = context.joinClassicGroupAndCompleteJoin(request, false, true);
         assertEquals(Errors.NONE.code(), response.errorCode());
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request.setMemberId("unknown-member-id")
         );
 
@@ -5531,7 +4408,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -5542,7 +4419,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData response = context.joinClassicGroupAsDynamicMemberAndCompleteJoin(request);
         assertEquals(Errors.NONE.code(), response.errorCode());
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request.setMemberId("other-member-id")
         );
 
@@ -5558,13 +4435,13 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
         group.transitionTo(DEAD);
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5576,13 +4453,13 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId("member-id")
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5597,25 +4474,25 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         // Leader joins
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertFalse(leaderJoinResult.joinFuture.isDone());
 
         // Member joins
-        JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(memberJoinResult.records.isEmpty());
         assertFalse(memberJoinResult.joinFuture.isDone());
 
         // Complete join group phase
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
         assertTrue(leaderJoinResult.joinFuture.isDone());
         assertTrue(memberJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), leaderJoinResult.joinFuture.get().errorCode());
@@ -5630,21 +4507,21 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2));
         assertFalse(joinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2 + 1));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2 + 1));
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
     }
@@ -5655,32 +4532,32 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .withRebalanceTimeoutMs(context.classicGroupInitialRebalanceDelayMs * 3)
             .build();
 
-        JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(firstMemberJoinResult.records.isEmpty());
         assertFalse(firstMemberJoinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs - 1));
-        JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs - 1));
+        GroupMetadataManagerTestContext.JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(secondMemberJoinResult.records.isEmpty());
         assertFalse(secondMemberJoinResult.joinFuture.isDone());
-        assertNoOrEmptyResult(context.sleep(2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2));
 
         // Advance clock past initial rebalance delay and verify futures are not completed.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2 + 1));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2 + 1));
         assertFalse(firstMemberJoinResult.joinFuture.isDone());
         assertFalse(secondMemberJoinResult.joinFuture.isDone());
 
         // Advance clock beyond recomputed delay and make sure the futures have completed.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs / 2));
         assertTrue(firstMemberJoinResult.joinFuture.isDone());
         assertTrue(secondMemberJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), firstMemberJoinResult.joinFuture.get().errorCode());
@@ -5693,38 +4570,38 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .withRebalanceTimeoutMs(context.classicGroupInitialRebalanceDelayMs * 2)
             .build();
 
-        JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(firstMemberJoinResult.records.isEmpty());
         assertFalse(firstMemberJoinResult.joinFuture.isDone());
 
-        JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(secondMemberJoinResult.records.isEmpty());
         assertFalse(secondMemberJoinResult.joinFuture.isDone());
 
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs + 1));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs + 1));
 
-        JoinResult thirdMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult thirdMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(thirdMemberJoinResult.records.isEmpty());
         assertFalse(thirdMemberJoinResult.joinFuture.isDone());
 
         // Advance clock right before rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs - 1));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs - 1));
         assertFalse(firstMemberJoinResult.joinFuture.isDone());
         assertFalse(secondMemberJoinResult.joinFuture.isDone());
         assertFalse(thirdMemberJoinResult.joinFuture.isDone());
 
         // Advance clock beyond rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(1));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(1));
         assertTrue(firstMemberJoinResult.joinFuture.isDone());
         assertTrue(secondMemberJoinResult.joinFuture.isDone());
         assertTrue(thirdMemberJoinResult.joinFuture.isDone());
@@ -5740,7 +4617,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withGroupInstanceId("group-instance-id")
@@ -5749,7 +4626,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         // Send join group as static member.
-        JoinResult oldMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult oldMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(oldMemberJoinResult.records.isEmpty());
         assertFalse(oldMemberJoinResult.joinFuture.isDone());
@@ -5757,7 +4634,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.size());
 
         // Replace static member with new member id. Old member id should be fenced.
-        JoinResult newMemberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult newMemberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(newMemberJoinResult.records.isEmpty());
         assertFalse(newMemberJoinResult.joinFuture.isDone());
@@ -5767,7 +4644,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.size());
 
         // Complete join for new member.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
         assertTrue(newMemberJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), newMemberJoinResult.joinFuture.get().errorCode());
         assertEquals(0, group.numAwaitingJoinResponse());
@@ -5780,14 +4657,14 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .withSessionTimeoutMs(1000)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5796,7 +4673,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.numPendingJoinMembers());
 
         // Advance clock by session timeout. Pending member should be removed from group as heartbeat expires.
-        assertNoOrEmptyResult(context.sleep(1000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(1000));
         assertEquals(0, group.numPendingJoinMembers());
     }
 
@@ -5808,13 +4685,13 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
@@ -5829,7 +4706,7 @@ public class GroupMetadataManagerTest {
         timeouts.forEach(timeout -> {
             assertEquals(classicGroupHeartbeatKey("group-id", memberId), timeout.key);
             assertEquals(Collections.singletonList(
-                newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+                GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
                 timeout.result.records());
         });
 
@@ -5844,7 +4721,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -5858,7 +4735,7 @@ public class GroupMetadataManagerTest {
 
         group.transitionTo(DEAD);
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -5871,13 +4748,13 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertEquals(Errors.MEMBER_ID_REQUIRED.code(), joinResult.joinFuture.get().errorCode());
@@ -5899,9 +4776,9 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestProtocolCollection protocols = toProtocols("range");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("range");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
@@ -5918,10 +4795,10 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(COMPLETING_REBALANCE));
         assertEquals(1, group.generationId());
 
-        protocols = toProtocols("range", "roundrobin");
+        protocols = GroupMetadataManagerTestContext.toProtocols("range", "roundrobin");
 
         // Send updated member metadata. This should trigger a rebalance and complete the join phase.
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request
                 .setMemberId(memberId)
                 .setProtocols(protocols)
@@ -5943,7 +4820,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -5960,7 +4837,7 @@ public class GroupMetadataManagerTest {
 
         group.transitionTo(STABLE);
         // Sending join group as leader should trigger a rebalance.
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request.setMemberId(memberId)
         );
 
@@ -5978,11 +4855,11 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
         JoinGroupResponseData leaderResponse = context.joinClassicGroupAsDynamicMemberAndCompleteJoin(request);
@@ -5991,13 +4868,13 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.generationId());
 
         // Member joins.
-        JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(memberJoinResult.records.isEmpty());
         assertFalse(memberJoinResult.joinFuture.isDone());
 
         // Leader also rejoins. Completes join group phase.
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(request.setMemberId(leaderId));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(request.setMemberId(leaderId));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -6013,7 +4890,7 @@ public class GroupMetadataManagerTest {
         // Member rejoins with updated metadata. This should trigger a rebalance.
         String memberId = memberJoinResult.joinFuture.get().memberId();
 
-        JoinGroupRequestProtocolCollection protocols = toProtocols("range", "roundrobin");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("range", "roundrobin");
 
         memberJoinResult = context.sendClassicGroupJoin(
             request
@@ -6042,7 +4919,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6054,13 +4931,13 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.generationId());
 
         // Member joins.
-        JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(memberJoinResult.records.isEmpty());
         assertFalse(memberJoinResult.joinFuture.isDone());
 
         // Leader also rejoins. Completes join group phase.
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(request.setMemberId(leaderId));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(request.setMemberId(leaderId));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -6090,7 +4967,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6106,7 +4983,7 @@ public class GroupMetadataManagerTest {
         group.transitionTo(PREPARING_REBALANCE);
         group.transitionTo(EMPTY);
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request.setMemberId(memberId));
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request.setMemberId(memberId));
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -6120,7 +4997,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6133,7 +5010,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.generationId());
 
         // Add new member. This triggers a rebalance.
-        JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult memberJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(memberJoinResult.records.isEmpty());
         assertFalse(memberJoinResult.joinFuture.isDone());
@@ -6141,7 +5018,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Advance clock by rebalance timeout. This will expire the leader as it has not rejoined.
-        assertNoOrEmptyResult(context.sleep(1000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(1000));
 
         assertTrue(memberJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), memberJoinResult.joinFuture.get().errorCode());
@@ -6156,7 +5033,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6164,7 +5041,7 @@ public class GroupMetadataManagerTest {
             .withRebalanceTimeoutMs(1000)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
@@ -6174,7 +5051,7 @@ public class GroupMetadataManagerTest {
         group.transitionTo(DEAD);
 
         // Advance clock by initial rebalance delay to complete join phase.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
         assertEquals(0, group.generationId());
     }
 
@@ -6184,7 +5061,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("first-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -6199,7 +5076,7 @@ public class GroupMetadataManagerTest {
         String firstMemberId = firstMemberResponse.memberId();
 
         // Second member joins and group goes into rebalancing state.
-        JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(
             request.setGroupInstanceId("second-instance-id")
         );
 
@@ -6207,7 +5084,7 @@ public class GroupMetadataManagerTest {
         assertFalse(secondMemberJoinResult.joinFuture.isDone());
 
         // First static member rejoins and completes join phase.
-        JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult firstMemberJoinResult = context.sendClassicGroupJoin(
             request.setMemberId(firstMemberId).setGroupInstanceId("first-instance-id"));
 
         assertTrue(firstMemberJoinResult.records.isEmpty());
@@ -6229,9 +5106,9 @@ public class GroupMetadataManagerTest {
 
         // Advance clock by rebalance timeout to complete join phase. As long as both members have not
         // rejoined, we extend the join phase.
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
         assertEquals(10000, context.timer.timeout("join-group-id").deadlineMs - context.time.milliseconds());
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
         assertEquals(10000, context.timer.timeout("join-group-id").deadlineMs - context.time.milliseconds());
 
         assertTrue(group.isInState(PREPARING_REBALANCE));
@@ -6276,12 +5153,12 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("group-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
         JoinGroupResponseData response = context.joinClassicGroupAndCompleteJoin(request, true, supportSkippingAssignment);
@@ -6296,9 +5173,9 @@ public class GroupMetadataManagerTest {
         group.transitionTo(STABLE);
 
         // Static member rejoins with UNKNOWN_MEMBER_ID. This should update the log with the generated member id.
-        JoinGroupRequestProtocolCollection protocols = toProtocols("range", "roundrobin");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("range", "roundrobin");
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request
                 .setProtocols(protocols)
                 .setRebalanceTimeoutMs(7000)
@@ -6308,7 +5185,7 @@ public class GroupMetadataManagerTest {
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             joinResult.records
         );
         assertFalse(joinResult.joinFuture.isDone());
@@ -6360,12 +5237,12 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("group-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
-            .withProtocols(toProtocols("range", "roundrobin"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range", "roundrobin"))
             .build();
 
         JoinGroupResponseData response = context.joinClassicGroupAndCompleteJoin(request, true, true);
@@ -6378,8 +5255,8 @@ public class GroupMetadataManagerTest {
         group.transitionTo(STABLE);
 
         // Static member rejoins with UNKNOWN_MEMBER_ID. The selected protocol changes and triggers a rebalance.
-        JoinResult joinResult = context.sendClassicGroupJoin(
-            request.setProtocols(toProtocols("roundrobin"))
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
+            request.setProtocols(GroupMetadataManagerTestContext.toProtocols("roundrobin"))
         );
 
         assertTrue(joinResult.records.isEmpty());
@@ -6397,9 +5274,9 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestProtocolCollection protocols = toProtocols("range");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("range");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("group-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -6425,7 +5302,7 @@ public class GroupMetadataManagerTest {
                 .setMetadata(ConsumerProtocol.serializeSubscription(new ConsumerPartitionAssignor.Subscription(
                     Collections.singletonList("bar"))).array()));
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request
                 .setProtocols(protocols)
                 .setRebalanceTimeoutMs(7000)
@@ -6435,7 +5312,7 @@ public class GroupMetadataManagerTest {
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             joinResult.records
         );
         assertFalse(joinResult.joinFuture.isDone());
@@ -6478,12 +5355,12 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("group-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolType("consumer")
-            .withProtocols(toProtocols("range"))
+            .withProtocols(GroupMetadataManagerTestContext.toProtocols("range"))
             .build();
 
         JoinGroupResponseData response = context.joinClassicGroupAndCompleteJoin(
@@ -6502,16 +5379,16 @@ public class GroupMetadataManagerTest {
         group.transitionTo(STABLE);
 
         // Static member rejoins with UNKNOWN_MEMBER_ID and the append succeeds.
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request
-                .setProtocols(toProtocols("range", "roundrobin"))
+                .setProtocols(GroupMetadataManagerTestContext.toProtocols("range", "roundrobin"))
                 .setRebalanceTimeoutMs(7000)
                 .setSessionTimeoutMs(6000),
             true,
             supportSkippingAssignment);
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             joinResult.records
         );
         assertFalse(joinResult.joinFuture.isDone());
@@ -6521,7 +5398,7 @@ public class GroupMetadataManagerTest {
         assertTrue(joinResult.joinFuture.isDone());
 
         JoinGroupResponseData expectedResponse = new JoinGroupResponseData()
-            .setMembers(supportSkippingAssignment ? toJoinResponseMembers(group) : Collections.emptyList())
+            .setMembers(supportSkippingAssignment ? GroupMetadataManagerTestContext.toJoinResponseMembers(group) : Collections.emptyList())
             .setLeader(supportSkippingAssignment ? joinResult.joinFuture.get().memberId() : oldMemberId)
             .setMemberId(joinResult.joinFuture.get().memberId())
             .setGenerationId(1)
@@ -6538,7 +5415,7 @@ public class GroupMetadataManagerTest {
         assertEquals(Optional.of("group-instance-id"), newMember.groupInstanceId());
         assertEquals(7000, newMember.rebalanceTimeoutMs());
         assertEquals(6000, newMember.sessionTimeoutMs());
-        assertEquals(toProtocols("range", "roundrobin"), newMember.supportedProtocols());
+        assertEquals(GroupMetadataManagerTestContext.toProtocols("range", "roundrobin"), newMember.supportedProtocols());
         assertEquals(1, group.size());
         assertEquals(1, group.generationId());
         assertTrue(group.isInState(STABLE));
@@ -6550,7 +5427,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("group-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -6565,7 +5442,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
         // Static member rejoins with UNKNOWN_MEMBER_ID and triggers a rebalance.
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -6597,8 +5474,8 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withDefaultProtocolTypeAndProtocols()
@@ -6610,7 +5487,7 @@ public class GroupMetadataManagerTest {
         assertFalse(joinResult.joinFuture.isDone());
 
         // Advance clock by initial rebalance delay to complete join phase.
-        assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupInitialRebalanceDelayMs));
 
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
@@ -6619,8 +5496,8 @@ public class GroupMetadataManagerTest {
 
         assertEquals(0, group.allMembers().stream().filter(ClassicGroupMember::isNew).count());
 
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(joinResult.joinFuture.get().memberId())
                 .withGenerationId(joinResult.joinFuture.get().generationId())
@@ -6635,12 +5512,12 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.size());
 
         // Make sure the NewMemberTimeout is not still in effect, and the member is not kicked
-        assertNoOrEmptyResult(context.sleep(context.classicGroupNewMemberJoinTimeoutMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(context.classicGroupNewMemberJoinTimeoutMs));
         assertEquals(1, group.size());
 
         // Member should be removed as heartbeat expires. The group is now empty.
         List<ExpiredTimeout<Void, Record>> timeouts = context.sleep(5000);
-        List<Record> expectedRecords = Collections.singletonList(newGroupMetadataRecord(
+        List<Record> expectedRecords = Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecord(
             group.groupId(),
             new GroupMetadataValue()
                 .setMembers(Collections.emptyList())
@@ -6673,7 +5550,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6688,8 +5565,8 @@ public class GroupMetadataManagerTest {
         assertEquals(memberId, joinResponse.leader());
         assertEquals(1, joinResponse.generationId());
 
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(memberId)
                 .withGenerationId(1)
@@ -6705,8 +5582,8 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(STABLE));
         assertEquals(1, group.generationId());
 
-        JoinResult otherJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(UNKNOWN_MEMBER_ID));
-        JoinResult joinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(memberId));
+        GroupMetadataManagerTestContext.JoinResult otherJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(UNKNOWN_MEMBER_ID));
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(memberId));
 
         assertTrue(otherJoinResult.records.isEmpty());
         assertTrue(joinResult.records.isEmpty());
@@ -6723,7 +5600,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -6731,7 +5608,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // A third member joins. Trigger a rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -6742,7 +5619,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Old follower rejoins group will be matching current member.id.
-        JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
             request
                 .setMemberId(rebalanceResult.followerId)
                 .setGroupInstanceId("follower-instance-id")
@@ -6770,7 +5647,7 @@ public class GroupMetadataManagerTest {
             .setMemberId(rebalanceResult.followerId)
             .setGenerationId(-1);
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             oldFollowerJoinResult.joinFuture.get(),
             group,
@@ -6784,7 +5661,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -6792,7 +5669,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Known leader rejoins will trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
@@ -6800,14 +5677,14 @@ public class GroupMetadataManagerTest {
             .withRebalanceTimeoutMs(10000)
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertFalse(leaderJoinResult.joinFuture.isDone());
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Old follower rejoins group will match current member.id.
-        JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
             request
                 .setMemberId(rebalanceResult.followerId)
                 .setGroupInstanceId("follower-instance-id")
@@ -6824,9 +5701,9 @@ public class GroupMetadataManagerTest {
             .setLeader(rebalanceResult.leaderId)
             .setProtocolName("range")
             .setProtocolType("consumer")
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedLeaderResponse,
             leaderJoinResult.joinFuture.get(),
             group,
@@ -6848,7 +5725,7 @@ public class GroupMetadataManagerTest {
             .setProtocolName("range")
             .setProtocolType("consumer");
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedFollowerResponse,
             oldFollowerJoinResult.joinFuture.get(),
             group,
@@ -6863,8 +5740,8 @@ public class GroupMetadataManagerTest {
         // Duplicate follower joins group with unknown member id will trigger member.id replacement,
         // and will also trigger a rebalance under CompletingRebalance state; the old follower sync callback
         // will return fenced exception while broker replaces the member identity with the duplicate follower joins.
-        SyncResult oldFollowerSyncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult oldFollowerSyncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("follower-instance-id")
                 .withGenerationId(oldFollowerJoinResult.joinFuture.get().generationId())
@@ -6875,7 +5752,7 @@ public class GroupMetadataManagerTest {
         assertTrue(oldFollowerSyncResult.records.isEmpty());
         assertFalse(oldFollowerSyncResult.syncFuture.isDone());
 
-        JoinResult duplicateFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult duplicateFollowerJoinResult = context.sendClassicGroupJoin(
             request
                 .setMemberId(UNKNOWN_MEMBER_ID)
                 .setGroupInstanceId("follower-instance-id")
@@ -6906,7 +5783,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -6914,7 +5791,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Known leader rejoins will trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
@@ -6922,14 +5799,14 @@ public class GroupMetadataManagerTest {
             .withRebalanceTimeoutMs(10000)
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertFalse(leaderJoinResult.joinFuture.isDone());
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Duplicate follower joins group will trigger member id replacement.
-        JoinResult duplicateFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult duplicateFollowerJoinResult = context.sendClassicGroupJoin(
             request
                 .setMemberId(UNKNOWN_MEMBER_ID)
                 .setGroupInstanceId("follower-instance-id")
@@ -6939,7 +5816,7 @@ public class GroupMetadataManagerTest {
         assertTrue(duplicateFollowerJoinResult.joinFuture.isDone());
 
         // Old follower rejoins group will fail because member id is already updated.
-        JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
             request.setMemberId(rebalanceResult.followerId)
         );
 
@@ -6954,9 +5831,9 @@ public class GroupMetadataManagerTest {
             .setLeader(rebalanceResult.leaderId)
             .setProtocolName("range")
             .setProtocolType("consumer")
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedLeaderResponse,
             leaderJoinResult.joinFuture.get(),
             group,
@@ -6975,7 +5852,7 @@ public class GroupMetadataManagerTest {
             .setProtocolType("consumer")
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedDuplicateFollowerResponse,
             duplicateFollowerJoinResult.joinFuture.get(),
             group,
@@ -6994,7 +5871,7 @@ public class GroupMetadataManagerTest {
             .setProtocolType(null)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedOldFollowerResponse,
             oldFollowerJoinResult.joinFuture.get(),
             group,
@@ -7009,7 +5886,7 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withGroupInstanceId("group-instance-id")
@@ -7021,7 +5898,7 @@ public class GroupMetadataManagerTest {
 
         String memberId = joinResponse.memberId();
 
-        JoinResult rejoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult rejoinResult = context.sendClassicGroupJoin(
             request.setMemberId(memberId)
         );
 
@@ -7031,8 +5908,8 @@ public class GroupMetadataManagerTest {
         assertTrue(rejoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), rejoinResult.joinFuture.get().errorCode());
 
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(memberId)
                 .withGenerationId(joinResponse.generationId())
@@ -7056,7 +5933,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7065,21 +5942,21 @@ public class GroupMetadataManagerTest {
 
         // A static leader rejoin with unknown id will not trigger rebalance, and no assignment will be returned.
         // As the group was in Stable state and the member id was updated, this will generate records.
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolSuperset()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             joinRequest,
             true,
             supportSkippingAssignment
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             joinResult.records
         );
         // Simulate a successful write to the log.
@@ -7090,7 +5967,7 @@ public class GroupMetadataManagerTest {
             joinResult.joinFuture.get().memberId() : rebalanceResult.leaderId;
 
         List<JoinGroupResponseMember> members = supportSkippingAssignment ?
-            toJoinResponseMembers(group) : Collections.emptyList();
+            GroupMetadataManagerTestContext.toJoinResponseMembers(group) : Collections.emptyList();
 
         JoinGroupResponseData expectedJoinResponse = new JoinGroupResponseData()
             .setErrorCode(Errors.NONE.code())
@@ -7102,7 +5979,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(supportSkippingAssignment)
             .setMembers(members);
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedJoinResponse,
             joinResult.joinFuture.get(),
             group,
@@ -7110,7 +5987,7 @@ public class GroupMetadataManagerTest {
             supportSkippingAssignment ? mkSet("leader-instance-id", "follower-instance-id") : Collections.emptySet()
         );
 
-        JoinResult oldLeaderJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult oldLeaderJoinResult = context.sendClassicGroupJoin(
             joinRequest.setMemberId(rebalanceResult.leaderId),
             true,
             supportSkippingAssignment
@@ -7121,14 +5998,14 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.FENCED_INSTANCE_ID.code(), oldLeaderJoinResult.joinFuture.get().errorCode());
 
         // Old leader will get fenced.
-        SyncGroupRequestData oldLeaderSyncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData oldLeaderSyncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withGenerationId(rebalanceResult.generationId)
             .withMemberId(rebalanceResult.leaderId)
             .build();
 
-        SyncResult oldLeaderSyncResult = context.sendClassicGroupSync(oldLeaderSyncRequest);
+        GroupMetadataManagerTestContext.SyncResult oldLeaderSyncResult = context.sendClassicGroupSync(oldLeaderSyncRequest);
 
         assertTrue(oldLeaderSyncResult.records.isEmpty());
         assertTrue(oldLeaderSyncResult.syncFuture.isDone());
@@ -7136,7 +6013,7 @@ public class GroupMetadataManagerTest {
 
         // Calling sync on old leader id will fail because that leader id is no longer valid and replaced.
         SyncGroupRequestData newLeaderSyncRequest = oldLeaderSyncRequest.setGroupInstanceId(null);
-        SyncResult newLeaderSyncResult = context.sendClassicGroupSync(newLeaderSyncRequest);
+        GroupMetadataManagerTestContext.SyncResult newLeaderSyncResult = context.sendClassicGroupSync(newLeaderSyncRequest);
 
         assertTrue(newLeaderSyncResult.records.isEmpty());
         assertTrue(newLeaderSyncResult.syncFuture.isDone());
@@ -7148,7 +6025,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7156,7 +6033,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Known static leader rejoin will trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
@@ -7176,9 +6053,9 @@ public class GroupMetadataManagerTest {
             .setLeader(rebalanceResult.leaderId)
             .setProtocolName("range")
             .setProtocolType("consumer")
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             joinResponse,
             group,
@@ -7192,7 +6069,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7200,14 +6077,14 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
         group.transitionTo(DEAD);
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -7219,7 +6096,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7229,14 +6106,14 @@ public class GroupMetadataManagerTest {
         group.transitionTo(PREPARING_REBALANCE);
         group.transitionTo(EMPTY);
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request, true, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request, true, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -7250,7 +6127,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id",
@@ -7260,8 +6137,8 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // A static follower rejoin with changed protocol will trigger rebalance.
-        JoinGroupRequestProtocolCollection protocols = toProtocols("roundrobin");
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("roundrobin");
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(rebalanceResult.followerId)
@@ -7270,13 +6147,13 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(sessionTimeoutMs)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
 
         // Old leader hasn't joined in the meantime, triggering a re-election.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
 
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
@@ -7290,9 +6167,9 @@ public class GroupMetadataManagerTest {
             .setLeader(rebalanceResult.followerId)
             .setProtocolName("roundrobin")
             .setProtocolType("consumer")
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             joinResult.joinFuture.get(),
             group,
@@ -7309,7 +6186,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id",
@@ -7321,9 +6198,9 @@ public class GroupMetadataManagerTest {
         assertNotEquals("roundrobin", group.selectProtocol());
 
         // A static follower rejoin with changed protocol will trigger rebalance.
-        JoinGroupRequestProtocolCollection protocols = toProtocols("roundrobin");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("roundrobin");
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -7332,13 +6209,13 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(sessionTimeoutMs)
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
         assertTrue(joinResult.records.isEmpty());
         assertFalse(joinResult.joinFuture.isDone());
         assertEquals("roundrobin", group.selectProtocol());
 
         // Old leader hasn't joined in the meantime, triggering a re-election.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
         assertTrue(group.hasStaticMember("leader-instance-id"));
@@ -7352,9 +6229,9 @@ public class GroupMetadataManagerTest {
             .setLeader(joinResult.joinFuture.get().memberId())
             .setProtocolName("roundrobin")
             .setProtocolType("consumer")
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             joinResult.joinFuture.get(),
             group,
@@ -7369,30 +6246,30 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
         );
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
-        JoinGroupRequestProtocolCollection protocols = toProtocols(group.selectProtocol());
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols(group.selectProtocol());
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocols(protocols)
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             followerJoinResult.records
         );
         // Simulate a failed write to the log.
@@ -7409,7 +6286,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7424,7 +6301,7 @@ public class GroupMetadataManagerTest {
         assertTrue(followerJoinResult.records.isEmpty());
 
         // Join with leader and complete join phase.
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(
             request.setGroupInstanceId("leader-instance-id")
                 .setMemberId(rebalanceResult.leaderId)
         );
@@ -7436,20 +6313,20 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.NONE.code(), followerJoinResult.joinFuture.get().errorCode());
 
         // Sync with leader and receive assignment.
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
             .withGenerationId(rebalanceResult.generationId + 1)
             .build();
 
-        SyncResult leaderSyncResult = context.sendClassicGroupSync(syncRequest);
+        GroupMetadataManagerTestContext.SyncResult leaderSyncResult = context.sendClassicGroupSync(syncRequest);
 
         // Simulate a successful write to the log. This will update the group with the new (empty) assignment.
         leaderSyncResult.appendFuture.complete(null);
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             leaderSyncResult.records
         );
 
@@ -7458,7 +6335,7 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.NONE.code(), leaderSyncResult.syncFuture.get().errorCode());
 
         // Sync with old member id will also not fail as the member id is not updated due to persistence failure
-        SyncResult oldMemberSyncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult oldMemberSyncResult = context.sendClassicGroupSync(
             syncRequest
                 .setGroupInstanceId("follower-instance-id")
                 .setMemberId(rebalanceResult.followerId)
@@ -7474,7 +6351,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7483,23 +6360,23 @@ public class GroupMetadataManagerTest {
 
         // A static follower rejoin with protocol changing to leader protocol subset won't trigger rebalance if updated
         // group's selectProtocol remain unchanged.
-        JoinGroupRequestProtocolCollection protocols = toProtocols(group.selectProtocol());
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols(group.selectProtocol());
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocols(protocols)
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             followerJoinResult.records
         );
 
@@ -7517,7 +6394,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7534,7 +6411,7 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.FENCED_INSTANCE_ID.code(), followerJoinResult.joinFuture.get().errorCode());
 
         // Sync with old member id will fail because the member id is updated
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withGenerationId(rebalanceResult.generationId)
@@ -7542,7 +6419,7 @@ public class GroupMetadataManagerTest {
             .withAssignment(Collections.emptyList())
             .build();
 
-        SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
 
         assertTrue(syncResult.records.isEmpty());
         assertTrue(syncResult.syncFuture.isDone());
@@ -7564,7 +6441,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7572,7 +6449,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // A static leader rejoin with known member id will trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.leaderId)
@@ -7581,7 +6458,7 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(5000)
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -7591,7 +6468,7 @@ public class GroupMetadataManagerTest {
         assertFalse(leaderJoinResult.joinFuture.isDone());
 
         // Rebalance completes immediately after follower rejoins.
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request.setGroupInstanceId("follower-instance-id")
                 .setMemberId(rebalanceResult.followerId),
             true,
@@ -7613,9 +6490,9 @@ public class GroupMetadataManagerTest {
             .setProtocolName("range")
             .setProtocolType("consumer")
             .setSkipAssignment(false)
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedLeaderResponse,
             leaderJoinResult.joinFuture.get(),
             group,
@@ -7633,7 +6510,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedFollowerResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7642,7 +6519,7 @@ public class GroupMetadataManagerTest {
         );
 
         // The follower protocol changed from protocolSuperset to general protocols.
-        JoinGroupRequestProtocolCollection protocols = toProtocols("range");
+        JoinGroupRequestProtocolCollection protocols = GroupMetadataManagerTestContext.toProtocols("range");
 
         followerJoinResult = context.sendClassicGroupJoin(
             request.setGroupInstanceId("follower-instance-id")
@@ -7673,9 +6550,9 @@ public class GroupMetadataManagerTest {
             .setProtocolName("range")
             .setProtocolType("consumer")
             .setSkipAssignment(false)
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedFollowerResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7689,7 +6566,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7697,21 +6574,21 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // A static follower rejoin with no protocol change will not trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolSuperset()
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
         );
 
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             followerJoinResult.records
         );
         // Simulate a successful write to log.
@@ -7732,7 +6609,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedFollowerResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7741,8 +6618,8 @@ public class GroupMetadataManagerTest {
         );
         assertNotEquals(rebalanceResult.followerId, followerJoinResult.joinFuture.get().memberId());
 
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("follower-instance-id")
                 .withGenerationId(rebalanceResult.generationId)
@@ -7761,7 +6638,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7769,14 +6646,14 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // A static follower rejoin with no protocol change will not trigger rebalance.
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(rebalanceResult.followerId)
             .withProtocolSuperset()
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -7799,7 +6676,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedFollowerResponse,
             followerJoinResult.joinFuture.get(),
             group,
@@ -7813,20 +6690,20 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
         );
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(rebalanceResult.followerId)
             .withProtocolSuperset()
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -7842,20 +6719,20 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
         );
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(rebalanceResult.leaderId)
             .withProtocolSuperset()
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -7877,13 +6754,13 @@ public class GroupMetadataManagerTest {
             "follower-instance-id"
         );
 
-        SyncGroupRequestData request = new SyncGroupRequestBuilder()
+        SyncGroupRequestData request = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId("invalid-member-id")
             .build();
 
-        SyncResult leaderSyncResult = context.sendClassicGroupSync(request);
+        GroupMetadataManagerTestContext.SyncResult leaderSyncResult = context.sendClassicGroupSync(request);
 
         assertTrue(leaderSyncResult.records.isEmpty());
         assertTrue(leaderSyncResult.syncFuture.isDone());
@@ -7895,7 +6772,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
@@ -7904,21 +6781,21 @@ public class GroupMetadataManagerTest {
 
         String lastMemberId = rebalanceResult.leaderId;
         for (int i = 0; i < 5; i++) {
-            JoinGroupRequestData request = new JoinGroupRequestBuilder()
+            JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("leader-instance-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withProtocolSuperset()
                 .build();
 
-            JoinResult leaderJoinResult = context.sendClassicGroupJoin(
+            GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(
                 request,
                 true,
                 true
             );
 
             assertEquals(
-                Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+                Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
                 leaderJoinResult.records
             );
             // Simulate a successful write to log.
@@ -7937,20 +6814,20 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
         );
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("unknown-instance-id")
             .withMemberId(rebalanceResult.leaderId)
             .withProtocolSuperset()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -7976,7 +6853,7 @@ public class GroupMetadataManagerTest {
         group.transitionTo(PREPARING_REBALANCE);
         group.transitionTo(EMPTY);
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("follower-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -8001,7 +6878,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         // Increase session timeout so that the follower won't be evicted when rebalance timeout is reached.
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id",
@@ -8013,14 +6890,14 @@ public class GroupMetadataManagerTest {
         String newMemberInstanceId = "new-member-instance-id";
         String leaderId = rebalanceResult.leaderId;
 
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId(newMemberInstanceId)
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolSuperset()
             .build();
 
-        JoinResult newMemberJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult newMemberJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -8030,7 +6907,7 @@ public class GroupMetadataManagerTest {
         assertFalse(newMemberJoinResult.joinFuture.isDone());
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(
             request
                 .setGroupInstanceId("leader-instance-id")
                 .setMemberId(leaderId),
@@ -8042,7 +6919,7 @@ public class GroupMetadataManagerTest {
         assertFalse(leaderJoinResult.joinFuture.isDone());
 
         // Advance clock by rebalance timeout to complete join phase.
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
 
         assertTrue(leaderJoinResult.joinFuture.isDone());
         assertTrue(newMemberJoinResult.joinFuture.isDone());
@@ -8055,9 +6932,9 @@ public class GroupMetadataManagerTest {
             .setProtocolName("range")
             .setProtocolType("consumer")
             .setSkipAssignment(false)
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedLeaderResponse,
             leaderJoinResult.joinFuture.get(),
             group,
@@ -8075,7 +6952,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedNewMemberResponse,
             newMemberJoinResult.joinFuture.get(),
             group,
@@ -8090,7 +6967,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         // Increase session timeout so that the leader won't be evicted when rebalance timeout is reached.
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id",
@@ -8100,14 +6977,14 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         String newMemberInstanceId = "new-member-instance-id";
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId(newMemberInstanceId)
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolSuperset()
             .build();
 
-        JoinResult newMemberJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult newMemberJoinResult = context.sendClassicGroupJoin(
             request,
             true,
             true
@@ -8117,7 +6994,7 @@ public class GroupMetadataManagerTest {
         assertFalse(newMemberJoinResult.joinFuture.isDone());
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
-        JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult oldFollowerJoinResult = context.sendClassicGroupJoin(
             request
                 .setGroupInstanceId("follower-instance-id")
                 .setMemberId(rebalanceResult.followerId),
@@ -8129,7 +7006,7 @@ public class GroupMetadataManagerTest {
         assertFalse(oldFollowerJoinResult.joinFuture.isDone());
 
         // Advance clock by rebalance timeout to complete join phase.
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
 
         assertTrue(oldFollowerJoinResult.joinFuture.isDone());
         assertTrue(newMemberJoinResult.joinFuture.isDone());
@@ -8151,9 +7028,9 @@ public class GroupMetadataManagerTest {
             .setProtocolName("range")
             .setProtocolType("consumer")
             .setSkipAssignment(false)
-            .setMembers(toJoinResponseMembers(group));
+            .setMembers(GroupMetadataManagerTestContext.toJoinResponseMembers(group));
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedLeaderResponse,
             newLeaderResponse,
             group,
@@ -8171,7 +7048,7 @@ public class GroupMetadataManagerTest {
             .setSkipAssignment(false)
             .setMembers(Collections.emptyList());
 
-        checkJoinGroupResponse(
+        GroupMetadataManagerTestContext.checkJoinGroupResponse(
             expectedNewMemberResponse,
             newFollowerResponse,
             group,
@@ -8236,26 +7113,26 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         // JoinGroup(leader) with the Protocol Type of the group
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withProtocolSuperset()
             .build();
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest);
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest);
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertFalse(leaderJoinResult.joinFuture.isDone());
 
         // JoinGroup(follower) with the Protocol Type of the group
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest.setGroupInstanceId("follower-instance-id"));
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest.setGroupInstanceId("follower-instance-id"));
 
         assertTrue(followerJoinResult.records.isEmpty());
         assertFalse(followerJoinResult.joinFuture.isDone());
 
         // Advance clock by rebalance timeout to complete join phase.
-        assertNoOrEmptyResult(context.sleep(10000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(10000));
         assertTrue(leaderJoinResult.joinFuture.isDone());
         assertTrue(followerJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), leaderJoinResult.joinFuture.get().errorCode());
@@ -8268,7 +7145,7 @@ public class GroupMetadataManagerTest {
         // SyncGroup with the provided Protocol Type and Name
         List<SyncGroupRequestAssignment> assignment = new ArrayList<>();
         assignment.add(new SyncGroupRequestAssignment().setMemberId(leaderId));
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(leaderId)
             .withProtocolType(protocolType.orElse(null))
@@ -8277,7 +7154,7 @@ public class GroupMetadataManagerTest {
             .withAssignment(assignment)
             .build();
 
-        SyncResult leaderSyncResult = context.sendClassicGroupSync(syncRequest);
+        GroupMetadataManagerTestContext.SyncResult leaderSyncResult = context.sendClassicGroupSync(syncRequest);
         // Simulate a successful write to the log.
         leaderSyncResult.appendFuture.complete(null);
 
@@ -8286,7 +7163,7 @@ public class GroupMetadataManagerTest {
         assertEquals(expectedProtocolType.orElse(null), leaderSyncResult.syncFuture.get().protocolType());
         assertEquals(expectedProtocolName.orElse(null), leaderSyncResult.syncFuture.get().protocolName());
 
-        SyncResult followerSyncResult = context.sendClassicGroupSync(syncRequest.setMemberId(followerId));
+        GroupMetadataManagerTestContext.SyncResult followerSyncResult = context.sendClassicGroupSync(syncRequest.setMemberId(followerId));
 
         assertTrue(followerSyncResult.records.isEmpty());
         assertTrue(followerSyncResult.syncFuture.isDone());
@@ -8301,8 +7178,8 @@ public class GroupMetadataManagerTest {
             .build();
 
         // SyncGroup with the provided Protocol Type and Name
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId("member-id")
                 .withGenerationId(1)
@@ -8321,7 +7198,7 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         JoinGroupResponseData joinResponse = context.joinClassicGroupAndCompleteJoin(
-            new JoinGroupRequestBuilder()
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("leader-instance-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
@@ -8336,14 +7213,14 @@ public class GroupMetadataManagerTest {
 
         List<SyncGroupRequestAssignment> assignment = new ArrayList<>();
         assignment.add(new SyncGroupRequestAssignment().setMemberId(memberId));
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(memberId)
             .withGenerationId(generationId)
             .withAssignment(assignment)
             .build();
 
-        SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
 
         // Simulate a successful write to log.
         syncResult.appendFuture.complete(null);
@@ -8367,7 +7244,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -8380,8 +7257,8 @@ public class GroupMetadataManagerTest {
         int generationId = joinResponse.generationId();
 
         // Send the sync group with an invalid generation
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(memberId)
                 .withGenerationId(generationId + 1)
@@ -8399,7 +7276,7 @@ public class GroupMetadataManagerTest {
             .build();
         context.createClassicGroup("group-id");
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -8409,8 +7286,8 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData joinResponse = context.joinClassicGroupAndCompleteJoin(joinRequest, true, true);
 
         // Send the sync group with an invalid generation
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(joinResponse.memberId())
                 .withGenerationId(1)
@@ -8434,18 +7311,18 @@ public class GroupMetadataManagerTest {
             .build();
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
 
         assertTrue(followerJoinResult.records.isEmpty());
         assertFalse(followerJoinResult.joinFuture.isDone());
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -8479,7 +7356,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -8487,12 +7364,12 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(5000)
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
 
         assertTrue(followerJoinResult.records.isEmpty());
         assertFalse(followerJoinResult.joinFuture.isDone());
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -8509,7 +7386,7 @@ public class GroupMetadataManagerTest {
 
         // With no leader SyncGroup, the follower's sync request should fail with an error indicating
         // that it should rejoin
-        SyncResult followerSyncResult = context.sendClassicGroupSync(new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult followerSyncResult = context.sendClassicGroupSync(new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(followerId)
             .withGenerationId(nextGenerationId)
@@ -8542,7 +7419,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -8550,12 +7427,12 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(5000)
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(UNKNOWN_MEMBER_ID));
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(UNKNOWN_MEMBER_ID));
 
         assertTrue(followerJoinResult.records.isEmpty());
         assertFalse(followerJoinResult.joinFuture.isDone());
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -8583,14 +7460,14 @@ public class GroupMetadataManagerTest {
             .setAssignment(followerAssignment)
         );
 
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(leaderJoinResponse.memberId())
             .withGenerationId(leaderJoinResponse.generationId())
             .withAssignment(assignment)
             .build();
 
-        SyncResult syncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
             syncRequest.setGenerationId(nextGenerationId)
         );
 
@@ -8602,7 +7479,7 @@ public class GroupMetadataManagerTest {
         assertEquals(leaderAssignment, syncResult.syncFuture.get().assignment());
 
         // Sync group with follower to get new assignment.
-        SyncResult followerSyncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult followerSyncResult = context.sendClassicGroupSync(
             syncRequest
                 .setMemberId(followerId)
                 .setGenerationId(nextGenerationId)
@@ -8626,7 +7503,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -8634,12 +7511,12 @@ public class GroupMetadataManagerTest {
             .withSessionTimeoutMs(5000)
             .build();
 
-        JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
+        GroupMetadataManagerTestContext.JoinResult followerJoinResult = context.sendClassicGroupJoin(joinRequest);
 
         assertTrue(followerJoinResult.records.isEmpty());
         assertFalse(followerJoinResult.joinFuture.isDone());
 
-        JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
+        GroupMetadataManagerTestContext.JoinResult leaderJoinResult = context.sendClassicGroupJoin(joinRequest.setMemberId(leaderJoinResponse.memberId()));
 
         assertTrue(leaderJoinResult.records.isEmpty());
         assertTrue(leaderJoinResult.joinFuture.isDone());
@@ -8657,13 +7534,13 @@ public class GroupMetadataManagerTest {
         byte[] followerAssignment = new byte[]{1};
 
         // Sync group with follower to get new assignment.
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(leaderJoinResponse.memberId())
             .withGenerationId(leaderJoinResponse.generationId())
             .build();
 
-        SyncResult followerSyncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult followerSyncResult = context.sendClassicGroupSync(
             syncRequest
                 .setMemberId(followerId)
                 .setGenerationId(nextGenerationId)
@@ -8683,7 +7560,7 @@ public class GroupMetadataManagerTest {
             .setAssignment(followerAssignment)
         );
 
-        SyncResult syncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
             syncRequest
                 .setMemberId(leaderJoinResponse.memberId())
                 .setGenerationId(nextGenerationId)
@@ -8721,8 +7598,8 @@ public class GroupMetadataManagerTest {
 
         // Join group from the leader should force the group to rebalance, which allows the
         // leader to push new assignment when local metadata changes
-        JoinResult leaderRejoinResult = context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        GroupMetadataManagerTestContext.JoinResult leaderRejoinResult = context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(leaderJoinResponse.memberId())
                 .withDefaultProtocolTypeAndProtocols()
@@ -8746,13 +7623,13 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData pendingMemberResponse = context.setupGroupWithPendingMember(group).pendingMemberResponse;
 
         // Compete join group for the pending member
-        JoinGroupRequestData request = new JoinGroupRequestBuilder()
+        JoinGroupRequestData request = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(pendingMemberResponse.memberId())
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(request);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(request);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -8775,7 +7652,7 @@ public class GroupMetadataManagerTest {
         // Advancing clock by > 2500 (session timeout for the third member)
         // and < 5000 (for first and second members). This will force the coordinator to attempt join
         // completion on heartbeat expiration (since we are in PendingRebalance stage).
-        assertNoOrEmptyResult(context.sleep(3000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(3000));
         assertTrue(group.isInState(COMPLETING_REBALANCE));
         assertEquals(2, group.allMembers().size());
         assertEquals(0, group.numPendingJoinMembers());
@@ -8787,8 +7664,8 @@ public class GroupMetadataManagerTest {
             .build();
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
 
-        JoinResult joinResult = context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(leaderJoinResponse.memberId())
                 .withDefaultProtocolTypeAndProtocols()
@@ -8806,14 +7683,14 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
 
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id"
         );
 
-        SyncResult syncResult = context.sendClassicGroupSync(
-            new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+            new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("leader-instance-id")
                 .withMemberId(rebalanceResult.leaderId)
@@ -8883,7 +7760,7 @@ public class GroupMetadataManagerTest {
             10000,
             5000,
             "consumer",
-            toProtocols("range")
+            GroupMetadataManagerTestContext.toProtocols("range")
         ));
 
         HeartbeatRequestData heartbeatRequest = new HeartbeatRequestData()
@@ -8915,13 +7792,13 @@ public class GroupMetadataManagerTest {
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
 
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
             .build();
 
-        JoinResult joinResult = context.sendClassicGroupJoin(joinRequest, true);
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(joinRequest, true);
 
         assertTrue(joinResult.records.isEmpty());
         assertTrue(joinResult.joinFuture.isDone());
@@ -8950,7 +7827,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
         JoinGroupResponseData leaderJoinResponse =
-            context.joinClassicGroupAsDynamicMemberAndCompleteJoin(new JoinGroupRequestBuilder()
+            context.joinClassicGroupAsDynamicMemberAndCompleteJoin(new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withDefaultProtocolTypeAndProtocols()
@@ -9026,7 +7903,7 @@ public class GroupMetadataManagerTest {
         JoinGroupResponseData leaderJoinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteRebalance("group-id");
 
         // Advance clock by 1/2 of session timeout.
-        assertNoOrEmptyResult(context.sleep(2500));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
 
         HeartbeatRequestData heartbeatRequest = new HeartbeatRequestData()
             .setGroupId("group-id")
@@ -9036,7 +7913,7 @@ public class GroupMetadataManagerTest {
         HeartbeatResponseData heartbeatResponse = context.sendClassicGroupHeartbeat(heartbeatRequest);
         assertEquals(Errors.NONE.code(), heartbeatResponse.errorCode());
 
-        assertNoOrEmptyResult(context.sleep(2500));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
 
         heartbeatResponse = context.sendClassicGroupHeartbeat(heartbeatRequest);
         assertEquals(Errors.NONE.code(), heartbeatResponse.errorCode());
@@ -9051,8 +7928,8 @@ public class GroupMetadataManagerTest {
 
         // Add a new member. This should trigger a rebalance. The new member has the
         // 'classicGroupNewMemberJoinTimeoutMs` session timeout, so it has a longer expiration than the existing member.
-        JoinResult otherJoinResult = context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        GroupMetadataManagerTestContext.JoinResult otherJoinResult = context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withDefaultProtocolTypeAndProtocols()
@@ -9066,7 +7943,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
         // Advance clock by 1/2 of session timeout.
-        assertNoOrEmptyResult(context.sleep(2500));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
 
         HeartbeatRequestData heartbeatRequest = new HeartbeatRequestData()
             .setGroupId("group-id")
@@ -9077,12 +7954,12 @@ public class GroupMetadataManagerTest {
         assertEquals(Errors.REBALANCE_IN_PROGRESS.code(), heartbeatResponse.errorCode());
 
         // Advance clock by first member's session timeout.
-        assertNoOrEmptyResult(context.sleep(5000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(5000));
 
         assertThrows(UnknownMemberIdException.class, () -> context.sendClassicGroupHeartbeat(heartbeatRequest));
 
         // Advance clock by remaining rebalance timeout to complete join phase.
-        assertNoOrEmptyResult(context.sleep(2500));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
 
         assertTrue(otherJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), otherJoinResult.joinFuture.get().errorCode());
@@ -9098,7 +7975,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
         // Create a group with a single member
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -9115,13 +7992,13 @@ public class GroupMetadataManagerTest {
         assertEquals(1, firstGenerationId);
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
-        SyncGroupRequestData syncRequest = new SyncGroupRequestBuilder()
+        SyncGroupRequestData syncRequest = new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(firstMemberId)
             .withGenerationId(firstGenerationId)
             .build();
 
-        SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(syncRequest);
 
         // Simulate a successful write to the log.
         syncResult.appendFuture.complete(null);
@@ -9132,7 +8009,7 @@ public class GroupMetadataManagerTest {
 
         // Add a new dynamic member. This should trigger a rebalance. The new member has the
         // 'classicGroupNewMemberJoinTimeoutMs` session timeout, so it has a longer expiration than the existing member.
-        JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult secondMemberJoinResult = context.sendClassicGroupJoin(
             joinRequest
                 .setMemberId(UNKNOWN_MEMBER_ID)
                 .setGroupInstanceId(null)
@@ -9150,7 +8027,7 @@ public class GroupMetadataManagerTest {
             .setGenerationId(firstGenerationId);
 
         for (int i = 0; i < 2; i++) {
-            assertNoOrEmptyResult(context.sleep(2500));
+            GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
             HeartbeatResponseData heartbeatResponse = context.sendClassicGroupHeartbeat(firstMemberHeartbeatRequest);
             assertEquals(Errors.REBALANCE_IN_PROGRESS.code(), heartbeatResponse.errorCode());
         }
@@ -9158,7 +8035,7 @@ public class GroupMetadataManagerTest {
         // Advance clock by remaining rebalance timeout to complete join phase.
         // The second member will become the leader. However, as the first member is a static member
         // it will not be kicked out.
-        assertNoOrEmptyResult(context.sleep(8000));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(8000));
 
         assertTrue(secondMemberJoinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), secondMemberJoinResult.joinFuture.get().errorCode());
@@ -9188,7 +8065,7 @@ public class GroupMetadataManagerTest {
         // Now session timeout the unjoined (first) member. Still keeping the new member.
         List<Errors> expectedErrors = Arrays.asList(Errors.NONE, Errors.NONE, Errors.REBALANCE_IN_PROGRESS);
         for (Errors expectedError : expectedErrors) {
-            assertNoOrEmptyResult(context.sleep(2000));
+            GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2000));
             HeartbeatResponseData heartbeatResponse = context.sendClassicGroupHeartbeat(
                 firstMemberHeartbeatRequest
                     .setMemberId(otherMemberId)
@@ -9200,7 +8077,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, group.size());
         assertTrue(group.isInState(PREPARING_REBALANCE));
 
-        JoinResult otherMemberRejoinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult otherMemberRejoinResult = context.sendClassicGroupJoin(
             joinRequest
                 .setMemberId(otherMemberId)
                 .setGroupInstanceId(null)
@@ -9213,7 +8090,7 @@ public class GroupMetadataManagerTest {
         assertEquals(3, otherMemberRejoinResult.joinFuture.get().generationId());
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
-        SyncResult otherMemberResyncResult = context.sendClassicGroupSync(
+        GroupMetadataManagerTestContext.SyncResult otherMemberResyncResult = context.sendClassicGroupSync(
             syncRequest
                 .setGroupInstanceId(null)
                 .setMemberId(otherMemberId)
@@ -9230,7 +8107,7 @@ public class GroupMetadataManagerTest {
         // The joined member should get heart beat response with no error. Let the new member keep
         // heartbeating for a while to verify that no new rebalance is triggered unexpectedly.
         for (int i = 0; i < 20; i++) {
-            assertNoOrEmptyResult(context.sleep(2000));
+            GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2000));
             HeartbeatResponseData heartbeatResponse = context.sendClassicGroupHeartbeat(
                 firstMemberHeartbeatRequest
                     .setMemberId(otherMemberId)
@@ -9268,7 +8145,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
         // Create a group with a single member
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withGroupInstanceId("leader-instance-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
@@ -9285,7 +8162,7 @@ public class GroupMetadataManagerTest {
         assertEquals(1, firstGenerationId);
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
-        SyncResult syncResult = context.sendClassicGroupSync(new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(firstMemberId)
             .withGenerationId(firstGenerationId)
@@ -9299,7 +8176,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(STABLE));
 
         // Add a new dynamic pending member.
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             joinRequest
                 .setMemberId(UNKNOWN_MEMBER_ID)
                 .setGroupInstanceId(null)
@@ -9321,7 +8198,7 @@ public class GroupMetadataManagerTest {
             .setGenerationId(firstGenerationId);
 
         for (int i = 0; i < 2; i++) {
-            assertNoOrEmptyResult(context.sleep(2500));
+            GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(2500));
             HeartbeatResponseData heartbeatResponse = context.sendClassicGroupHeartbeat(heartbeatRequest);
             assertEquals(Errors.NONE.code(), heartbeatResponse.errorCode());
         }
@@ -9347,7 +8224,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Advance clock by 1/2 rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // Heartbeats to ensure that heartbeating does not interfere with the
         // delayed sync operation.
@@ -9360,7 +8237,7 @@ public class GroupMetadataManagerTest {
         ExpiredTimeout<Void, Record> timeout = timeouts.get(0);
         assertEquals(classicGroupSyncKey("group-id"), timeout.key);
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             timeout.result.records()
         );
 
@@ -9385,14 +8262,14 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Advance clock by 1/2 rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // Heartbeats to ensure that heartbeating does not interfere with the
         // delayed sync operation.
         joinResponses.forEach(response -> context.verifyHeartbeat(group.groupId(), response, Errors.NONE));
 
         // Leader sends a sync group request.
-        SyncResult syncResult = context.sendClassicGroupSync(new SyncGroupRequestBuilder()
+        GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
             .withGroupId("group-id")
             .withGenerationId(1)
             .withMemberId(joinResponses.get(0).memberId())
@@ -9437,7 +8314,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // Advance clock by 1/2 rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // Heartbeats to ensure that heartbeating does not interfere with the
         // delayed sync operation.
@@ -9446,8 +8323,8 @@ public class GroupMetadataManagerTest {
         // Followers send sync group requests.
         List<CompletableFuture<SyncGroupResponseData>> followerSyncFutures = joinResponses.subList(1, 3).stream()
             .map(response -> {
-                SyncResult syncResult = context.sendClassicGroupSync(
-                    new SyncGroupRequestBuilder()
+                GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+                    new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                         .withGroupId("group-id")
                         .withGenerationId(1)
                         .withMemberId(response.memberId())
@@ -9498,7 +8375,7 @@ public class GroupMetadataManagerTest {
         String leaderId = joinResponses.get(0).memberId();
 
         // Advance clock by 1/2 rebalance timeout.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // Heartbeats to ensure that heartbeating does not interfere with the
         // delayed sync operation.
@@ -9506,8 +8383,8 @@ public class GroupMetadataManagerTest {
 
         // All members send sync group requests.
         List<CompletableFuture<SyncGroupResponseData>> syncFutures = joinResponses.stream().map(response -> {
-            SyncResult syncResult = context.sendClassicGroupSync(
-                new SyncGroupRequestBuilder()
+            GroupMetadataManagerTestContext.SyncResult syncResult = context.sendClassicGroupSync(
+                new GroupMetadataManagerTestContext.SyncGroupRequestBuilder()
                     .withGroupId("group-id")
                     .withGenerationId(1)
                     .withMemberId(response.memberId())
@@ -9516,7 +8393,7 @@ public class GroupMetadataManagerTest {
 
             if (response.memberId().equals(leaderId)) {
                 assertEquals(
-                    Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+                    Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
                     syncResult.records
                 );
 
@@ -9534,13 +8411,13 @@ public class GroupMetadataManagerTest {
         }
 
         // Advance clock by 1/2 rebalance timeout. Pending sync should already have been cancelled.
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // All member heartbeats should succeed.
         joinResponses.forEach(response -> context.verifyHeartbeat(group.groupId(), response, Errors.NONE));
 
         // Advance clock a bit more
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs / 2));
 
         // All member heartbeats should succeed.
         joinResponses.forEach(response -> context.verifyHeartbeat(group.groupId(), response, Errors.NONE));
@@ -9555,7 +8432,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
         // First start up a group (with a slightly larger timeout to give us time to heartbeat when the rebalance starts)
-        JoinGroupRequestData joinRequest = new JoinGroupRequestBuilder()
+        JoinGroupRequestData joinRequest = new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
             .withGroupId("group-id")
             .withMemberId(UNKNOWN_MEMBER_ID)
             .withDefaultProtocolTypeAndProtocols()
@@ -9570,7 +8447,7 @@ public class GroupMetadataManagerTest {
         assertTrue(group.isInState(COMPLETING_REBALANCE));
 
         // Then join with a new consumer to trigger a rebalance
-        JoinResult joinResult = context.sendClassicGroupJoin(
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
             joinRequest.setMemberId(UNKNOWN_MEMBER_ID)
         );
 
@@ -9601,7 +8478,7 @@ public class GroupMetadataManagerTest {
             .build();
 
         // Create one classic group record.
-        context.replay(newGroupMetadataRecord(
+        context.replay(GroupMetadataManagerTestContext.newGroupMetadataRecord(
             classicGroupId,
             new GroupMetadataValue()
                 .setMembers(Collections.emptyList())
@@ -9872,7 +8749,7 @@ public class GroupMetadataManagerTest {
             .setProtocol("range")
             .setCurrentStateTimestamp(context.time.milliseconds());
 
-        context.replay(newGroupMetadataRecord(
+        context.replay(GroupMetadataManagerTestContext.newGroupMetadataRecord(
             "group-id",
             groupMetadataValue,
             MetadataVersion.latestTesting()
@@ -9921,7 +8798,7 @@ public class GroupMetadataManagerTest {
             .setProtocol("range")
             .setCurrentStateTimestamp(context.time.milliseconds());
 
-        context.replay(newGroupMetadataRecord(
+        context.replay(GroupMetadataManagerTestContext.newGroupMetadataRecord(
             "group-id",
             groupMetadataValue,
             MetadataVersion.latestTesting()
@@ -9961,182 +8838,13 @@ public class GroupMetadataManagerTest {
         context.verifyDescribeGroupsReturnsDeadGroup("group-id");
     }
 
-    public static <T> void assertUnorderedListEquals(
-        List<T> expected,
-        List<T> actual
-    ) {
-        assertEquals(new HashSet<>(expected), new HashSet<>(actual));
-    }
-
-    private static void assertResponseEquals(
-        ConsumerGroupHeartbeatResponseData expected,
-        ConsumerGroupHeartbeatResponseData actual
-    ) {
-        if (!responseEquals(expected, actual)) {
-            assertionFailure()
-                .expected(expected)
-                .actual(actual)
-                .buildAndThrow();
-        }
-    }
-
-    private static boolean responseEquals(
-        ConsumerGroupHeartbeatResponseData expected,
-        ConsumerGroupHeartbeatResponseData actual
-    ) {
-        if (expected.throttleTimeMs() != actual.throttleTimeMs()) return false;
-        if (expected.errorCode() != actual.errorCode()) return false;
-        if (!Objects.equals(expected.errorMessage(), actual.errorMessage())) return false;
-        if (!Objects.equals(expected.memberId(), actual.memberId())) return false;
-        if (expected.memberEpoch() != actual.memberEpoch()) return false;
-        if (expected.heartbeatIntervalMs() != actual.heartbeatIntervalMs()) return false;
-        // Unordered comparison of the assignments.
-        return responseAssignmentEquals(expected.assignment(), actual.assignment());
-    }
-
-    private static boolean responseAssignmentEquals(
-        ConsumerGroupHeartbeatResponseData.Assignment expected,
-        ConsumerGroupHeartbeatResponseData.Assignment actual
-    ) {
-        if (expected == actual) return true;
-        if (expected == null) return false;
-        if (actual == null) return false;
-
-        return Objects.equals(fromAssignment(expected.topicPartitions()), fromAssignment(actual.topicPartitions()));
-    }
-
-    private static Map<Uuid, Set<Integer>> fromAssignment(
-        List<ConsumerGroupHeartbeatResponseData.TopicPartitions> assignment
-    ) {
-        if (assignment == null) return null;
-
-        Map<Uuid, Set<Integer>> assignmentMap = new HashMap<>();
-        assignment.forEach(topicPartitions -> {
-            assignmentMap.put(topicPartitions.topicId(), new HashSet<>(topicPartitions.partitions()));
-        });
-        return assignmentMap;
-    }
-
-    private static void assertRecordsEquals(
-        List<Record> expectedRecords,
-        List<Record> actualRecords
-    ) {
-        try {
-            assertEquals(expectedRecords.size(), actualRecords.size());
-
-            for (int i = 0; i < expectedRecords.size(); i++) {
-                Record expectedRecord = expectedRecords.get(i);
-                Record actualRecord = actualRecords.get(i);
-                assertRecordEquals(expectedRecord, actualRecord);
-            }
-        } catch (AssertionFailedError e) {
-            assertionFailure()
-                .expected(expectedRecords)
-                .actual(actualRecords)
-                .buildAndThrow();
-        }
-    }
-
-    private static void assertRecordEquals(
-        Record expected,
-        Record actual
-    ) {
-        try {
-            assertApiMessageAndVersionEquals(expected.key(), actual.key());
-            assertApiMessageAndVersionEquals(expected.value(), actual.value());
-        } catch (AssertionFailedError e) {
-            assertionFailure()
-                .expected(expected)
-                .actual(actual)
-                .buildAndThrow();
-        }
-    }
-
-    private static void assertApiMessageAndVersionEquals(
-        ApiMessageAndVersion expected,
-        ApiMessageAndVersion actual
-    ) {
-        if (expected == actual) return;
-
-        assertEquals(expected.version(), actual.version());
-
-        if (actual.message() instanceof ConsumerGroupCurrentMemberAssignmentValue) {
-            // The order of the topics stored in ConsumerGroupCurrentMemberAssignmentValue is not
-            // always guaranteed. Therefore, we need a special comparator.
-            ConsumerGroupCurrentMemberAssignmentValue expectedValue =
-                (ConsumerGroupCurrentMemberAssignmentValue) expected.message();
-            ConsumerGroupCurrentMemberAssignmentValue actualValue =
-                (ConsumerGroupCurrentMemberAssignmentValue) actual.message();
-
-            assertEquals(expectedValue.memberEpoch(), actualValue.memberEpoch());
-            assertEquals(expectedValue.previousMemberEpoch(), actualValue.previousMemberEpoch());
-            assertEquals(expectedValue.targetMemberEpoch(), actualValue.targetMemberEpoch());
-            assertEquals(expectedValue.error(), actualValue.error());
-            assertEquals(expectedValue.metadataVersion(), actualValue.metadataVersion());
-            assertEquals(expectedValue.metadataBytes(), actualValue.metadataBytes());
-
-            // We transform those to Maps before comparing them.
-            assertEquals(fromTopicPartitions(expectedValue.assignedPartitions()),
-                fromTopicPartitions(actualValue.assignedPartitions()));
-            assertEquals(fromTopicPartitions(expectedValue.partitionsPendingRevocation()),
-                fromTopicPartitions(actualValue.partitionsPendingRevocation()));
-            assertEquals(fromTopicPartitions(expectedValue.partitionsPendingAssignment()),
-                fromTopicPartitions(actualValue.partitionsPendingAssignment()));
-        } else {
-            assertEquals(expected.message(), actual.message());
-        }
-    }
-
-    private static Map<Uuid, Set<Integer>> fromTopicPartitions(
-        List<ConsumerGroupCurrentMemberAssignmentValue.TopicPartitions> assignment
-    ) {
-        Map<Uuid, Set<Integer>> assignmentMap = new HashMap<>();
-        assignment.forEach(topicPartitions -> {
-            assignmentMap.put(topicPartitions.topicId(), new HashSet<>(topicPartitions.partitions()));
-        });
-        return assignmentMap;
-    }
-
-    private static List<JoinGroupResponseMember> toJoinResponseMembers(ClassicGroup group) {
-        List<JoinGroupResponseMember> members = new ArrayList<>();
-        String protocolName = group.protocolName().get();
-        group.allMembers().forEach(member -> {
-            members.add(
-                new JoinGroupResponseMember()
-                    .setMemberId(member.memberId())
-                    .setGroupInstanceId(member.groupInstanceId().orElse(""))
-                    .setMetadata(member.metadata(protocolName))
-            );
-        });
-
-        return members;
-    }
-
-    private static void checkJoinGroupResponse(
-        JoinGroupResponseData expectedResponse,
-        JoinGroupResponseData actualResponse,
-        ClassicGroup group,
-        ClassicGroupState expectedState,
-        Set<String> expectedGroupInstanceIds
-    ) {
-        assertEquals(expectedResponse, actualResponse);
-        assertTrue(group.isInState(expectedState));
-
-        Set<String> groupInstanceIds = actualResponse.members()
-            .stream()
-            .map(JoinGroupResponseData.JoinGroupResponseMember::groupInstanceId)
-            .collect(Collectors.toSet());
-
-        assertEquals(expectedGroupInstanceIds, groupInstanceIds);
-    }
-
     @Test
     public void testGroupStuckInRebalanceTimeoutDueToNonjoinedStaticMember() throws Exception {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
         int longSessionTimeoutMs = 10000;
         int rebalanceTimeoutMs = 5000;
-        RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
+        GroupMetadataManagerTestContext.RebalanceResult rebalanceResult = context.staticMembersJoinAndRebalance(
             "group-id",
             "leader-instance-id",
             "follower-instance-id",
@@ -10146,8 +8854,8 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.groupMetadataManager.getOrMaybeCreateClassicGroup("group-id", false);
 
         // New member joins
-        JoinResult joinResult = context.sendClassicGroupJoin(
-            new JoinGroupRequestBuilder()
+        GroupMetadataManagerTestContext.JoinResult joinResult = context.sendClassicGroupJoin(
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withProtocolSuperset()
@@ -10156,7 +8864,7 @@ public class GroupMetadataManagerTest {
         );
 
         // The new dynamic member has been elected as leader
-        assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
+        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(rebalanceTimeoutMs));
         assertTrue(joinResult.joinFuture.isDone());
         assertEquals(Errors.NONE.code(), joinResult.joinFuture.get().errorCode());
         assertEquals(joinResult.joinFuture.get().leader(), joinResult.joinFuture.get().memberId());
@@ -10271,7 +8979,7 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         context.joinClassicGroupAsDynamicMemberAndCompleteJoin(
-            new JoinGroupRequestBuilder()
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withDefaultProtocolTypeAndProtocols()
@@ -10328,7 +9036,7 @@ public class GroupMetadataManagerTest {
         ClassicGroup group = context.createClassicGroup("group-id");
 
         JoinGroupResponseData joinResponse = context.joinClassicGroupAsDynamicMemberAndCompleteJoin(
-            new JoinGroupRequestBuilder()
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
                 .withDefaultProtocolTypeAndProtocols()
@@ -10345,7 +9053,7 @@ public class GroupMetadataManagerTest {
                 ))
         );
         assertEquals(
-            Collections.singletonList(newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
+            Collections.singletonList(GroupMetadataManagerTestContext.newGroupMetadataRecordWithCurrentState(group, MetadataVersion.latestTesting())),
             leaveResult.records()
         );
         // Simulate a successful write to the log.
@@ -10369,7 +9077,7 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         context.joinClassicGroupAndCompleteJoin(
-            new JoinGroupRequestBuilder()
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("group-instance-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
@@ -10407,7 +9115,7 @@ public class GroupMetadataManagerTest {
         context.createClassicGroup("group-id");
 
         context.joinClassicGroupAndCompleteJoin(
-            new JoinGroupRequestBuilder()
+            new GroupMetadataManagerTestContext.JoinGroupRequestBuilder()
                 .withGroupId("group-id")
                 .withGroupInstanceId("group-instance-id")
                 .withMemberId(UNKNOWN_MEMBER_ID)
@@ -10601,7 +9309,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
-        PendingMemberGroupResult pendingMemberGroupResult = context.setupGroupWithPendingMember(group);
+        GroupMetadataManagerTestContext.PendingMemberGroupResult pendingMemberGroupResult = context.setupGroupWithPendingMember(group);
 
         CoordinatorResult<LeaveGroupResponseData, Record> leaveResult = context.sendClassicGroupLeave(
             new LeaveGroupRequestData()
@@ -10638,7 +9346,7 @@ public class GroupMetadataManagerTest {
         GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
             .build();
         ClassicGroup group = context.createClassicGroup("group-id");
-        PendingMemberGroupResult pendingMemberGroupResult = context.setupGroupWithPendingMember(group);
+        GroupMetadataManagerTestContext.PendingMemberGroupResult pendingMemberGroupResult = context.setupGroupWithPendingMember(group);
 
         CoordinatorResult<LeaveGroupResponseData, Record> leaveResult = context.sendClassicGroupLeave(
             new LeaveGroupRequestData()
@@ -10895,274 +9603,5 @@ public class GroupMetadataManagerTest {
 
         verify(context.metrics, times(1)).onConsumerGroupStateTransition(null, ConsumerGroup.ConsumerGroupState.EMPTY);
         verify(context.metrics, times(1)).onConsumerGroupStateTransition(ConsumerGroup.ConsumerGroupState.EMPTY, null);
-    }
-
-    private static void assertNoOrEmptyResult(List<ExpiredTimeout<Void, Record>> timeouts) {
-        assertTrue(timeouts.size() <= 1);
-        timeouts.forEach(timeout -> assertEquals(EMPTY_RESULT, timeout.result));
-    }
-
-    private static List<String> verifyClassicGroupJoinResponses(
-        List<JoinResult> joinResults,
-        int expectedSuccessCount,
-        Errors expectedFailure
-    ) {
-        int successCount = 0;
-        List<String> memberIds = new ArrayList<>();
-        for (JoinResult joinResult : joinResults) {
-            if (!joinResult.joinFuture.isDone()) {
-                fail("All responseFutures should be completed.");
-            }
-            try {
-                JoinGroupResponseData joinResponse = joinResult.joinFuture.get();
-                if (joinResponse.errorCode() == Errors.NONE.code()) {
-                    successCount++;
-                } else {
-                    assertEquals(expectedFailure.code(), joinResponse.errorCode());
-                }
-                memberIds.add(joinResponse.memberId());
-            } catch (Exception e) {
-                fail("Unexpected exception: " + e.getMessage());
-            }
-        }
-
-        assertEquals(expectedSuccessCount, successCount);
-        return memberIds;
-    }
-
-    private static JoinGroupRequestProtocolCollection toProtocols(String... protocolNames) {
-        JoinGroupRequestProtocolCollection protocols = new JoinGroupRequestProtocolCollection(0);
-        List<String> topicNames = Arrays.asList("foo", "bar", "baz");
-        for (int i = 0; i < protocolNames.length; i++) {
-            protocols.add(new JoinGroupRequestProtocol()
-                .setName(protocolNames[i])
-                .setMetadata(ConsumerProtocol.serializeSubscription(new ConsumerPartitionAssignor.Subscription(
-                    Collections.singletonList(topicNames.get(i % topicNames.size())))).array())
-            );
-        }
-        return protocols;
-    }
-
-    private static Record newGroupMetadataRecord(
-        String groupId,
-        GroupMetadataValue value,
-        MetadataVersion metadataVersion
-    ) {
-        return new Record(
-            new ApiMessageAndVersion(
-                new GroupMetadataKey()
-                    .setGroup(groupId),
-                (short) 2
-            ),
-            new ApiMessageAndVersion(
-                value,
-                metadataVersion.groupMetadataValueVersion()
-            )
-        );
-    }
-
-    private static Record newGroupMetadataRecordWithCurrentState(
-        ClassicGroup group,
-        MetadataVersion metadataVersion
-    ) {
-        return RecordHelpers.newGroupMetadataRecord(group, group.groupAssignment(), metadataVersion);
-    }
-
-    private static class JoinGroupRequestBuilder {
-        String groupId = null;
-        String groupInstanceId = null;
-        String memberId = null;
-        String protocolType = "consumer";
-        JoinGroupRequestProtocolCollection protocols = new JoinGroupRequestProtocolCollection(0);
-        int sessionTimeoutMs = 500;
-        int rebalanceTimeoutMs = 500;
-        String reason = null;
-
-        JoinGroupRequestBuilder withGroupId(String groupId) {
-            this.groupId = groupId;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withGroupInstanceId(String groupInstanceId) {
-            this.groupInstanceId = groupInstanceId;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withMemberId(String memberId) {
-            this.memberId = memberId;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withDefaultProtocolTypeAndProtocols() {
-            this.protocols = toProtocols("range");
-            return this;
-        }
-
-        JoinGroupRequestBuilder withProtocolSuperset() {
-            this.protocols = toProtocols("range", "roundrobin");
-            return this;
-        }
-
-        JoinGroupRequestBuilder withProtocolType(String protocolType) {
-            this.protocolType = protocolType;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withProtocols(JoinGroupRequestProtocolCollection protocols) {
-            this.protocols = protocols;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withRebalanceTimeoutMs(int rebalanceTimeoutMs) {
-            this.rebalanceTimeoutMs = rebalanceTimeoutMs;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withSessionTimeoutMs(int sessionTimeoutMs) {
-            this.sessionTimeoutMs = sessionTimeoutMs;
-            return this;
-        }
-
-        JoinGroupRequestBuilder withReason(String reason) {
-            this.reason = reason;
-            return this;
-        }
-
-        JoinGroupRequestData build() {
-            return new JoinGroupRequestData()
-                .setGroupId(groupId)
-                .setGroupInstanceId(groupInstanceId)
-                .setMemberId(memberId)
-                .setProtocolType(protocolType)
-                .setProtocols(protocols)
-                .setRebalanceTimeoutMs(rebalanceTimeoutMs)
-                .setSessionTimeoutMs(sessionTimeoutMs)
-                .setReason(reason);
-        }
-    }
-
-    private static class SyncGroupRequestBuilder {
-        String groupId = null;
-        String groupInstanceId = null;
-        String memberId = null;
-        String protocolType = "consumer";
-        String protocolName = "range";
-        int generationId = 0;
-        List<SyncGroupRequestAssignment> assignment = Collections.emptyList();
-
-        SyncGroupRequestBuilder withGroupId(String groupId) {
-            this.groupId = groupId;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withGroupInstanceId(String groupInstanceId) {
-            this.groupInstanceId = groupInstanceId;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withMemberId(String memberId) {
-            this.memberId = memberId;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withGenerationId(int generationId) {
-            this.generationId = generationId;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withProtocolType(String protocolType) {
-            this.protocolType = protocolType;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withProtocolName(String protocolName) {
-            this.protocolName = protocolName;
-            return this;
-        }
-
-        SyncGroupRequestBuilder withAssignment(List<SyncGroupRequestAssignment> assignment) {
-            this.assignment = assignment;
-            return this;
-        }
-
-
-        SyncGroupRequestData build() {
-            return new SyncGroupRequestData()
-                .setGroupId(groupId)
-                .setGroupInstanceId(groupInstanceId)
-                .setMemberId(memberId)
-                .setGenerationId(generationId)
-                .setProtocolType(protocolType)
-                .setProtocolName(protocolName)
-                .setAssignments(assignment);
-        }
-    }
-
-    private static class RebalanceResult {
-        int generationId;
-        String leaderId;
-        byte[] leaderAssignment;
-        String followerId;
-        byte[] followerAssignment;
-
-        RebalanceResult(
-            int generationId,
-            String leaderId,
-            byte[] leaderAssignment,
-            String followerId,
-            byte[] followerAssignment
-        ) {
-            this.generationId = generationId;
-            this.leaderId = leaderId;
-            this.leaderAssignment = leaderAssignment;
-            this.followerId = followerId;
-            this.followerAssignment = followerAssignment;
-        }
-    }
-
-    private static class PendingMemberGroupResult {
-        String leaderId;
-        String followerId;
-        JoinGroupResponseData pendingMemberResponse;
-
-        public PendingMemberGroupResult(
-            String leaderId,
-            String followerId,
-            JoinGroupResponseData pendingMemberResponse
-        ) {
-            this.leaderId = leaderId;
-            this.followerId = followerId;
-            this.pendingMemberResponse = pendingMemberResponse;
-        }
-    }
-
-    private static class JoinResult {
-        CompletableFuture<JoinGroupResponseData> joinFuture;
-        List<Record> records;
-        CompletableFuture<Void> appendFuture;
-
-        public JoinResult(
-            CompletableFuture<JoinGroupResponseData> joinFuture,
-            CoordinatorResult<Void, Record> coordinatorResult
-        ) {
-            this.joinFuture = joinFuture;
-            this.records = coordinatorResult.records();
-            this.appendFuture = coordinatorResult.appendFuture();
-        }
-    }
-
-    private static class SyncResult {
-        CompletableFuture<SyncGroupResponseData> syncFuture;
-        List<Record> records;
-        CompletableFuture<Void> appendFuture;
-
-        public SyncResult(
-            CompletableFuture<SyncGroupResponseData> syncFuture,
-            CoordinatorResult<Void, Record> coordinatorResult
-        ) {
-            this.syncFuture = syncFuture;
-            this.records = coordinatorResult.records();
-            this.appendFuture = coordinatorResult.appendFuture();
-        }
     }
 }
